The implicity register conversion between 32-bit and 64-bit does not work.
Add explicit x86 32-bit register conversion from Capstone to Dyninst.

Split Capstone x86 code into a separate file to reduce compilation time
