+++ instructionAPI/CMakeLists.txt
@@ -23,6 +23,8 @@ set (SRC_LIST
      src/InstructionDecoder-amdgpu-vega.C
      src/InstructionDecoderImpl.C
      src/InstructionDecoder-Capstone.C
+     src/InstructionDecoder-Capstone-x86.C
+
   )
 SET_SOURCE_FILES_PROPERTIES(${SRC_LIST} PROPERTIES LANGUAGE CXX)
