+++ instructionAPI/src/InstructionDecoder-Capstone.C
@@ -1702,3671 +1702,6 @@ void InstructionDecoder_Capstone::initializeOpcodeData() {
 }
 
 
-/******************************  x86 functions  *************************************/
-
-void InstructionDecoder_Capstone::decodeOperands_x86(const Instruction* insn, cs_detail *d) {
-    bool isCFT = false;
-    bool isCall = false;
-    bool isConditional = false;
-    InsnCategory cat = insn->getCategory();
-     if (cat == c_ReturnInsn) {
-        Expression::Ptr ret_addr = makeDereferenceExpression(makeRegisterExpression(x86_64::rsp), u64);
-        insn->addSuccessor(ret_addr, false, true, false, false);
-        return;
-    }
-    if(cat == c_BranchInsn || cat == c_CallInsn) {
-        isCFT = true;
-        if(cat == c_CallInsn) {
-            isCall = true;
-        }
-    }
-
-    if(cat == c_BranchInsn && insn->getOperation().getID() != e_jmp) {
-        isConditional = true;
-    }
-    bool isRela = checkCapstoneGroup(d, (uint8_t)CS_GRP_BRANCH_RELATIVE);
-    bool err = false;
-    cs_x86* detail = &(d->x86);
-    for (uint8_t i = 0; i < detail->op_count; ++i) {
-        cs_x86_op* operand = &(detail->operands[i]);
-        if (operand->type == X86_OP_REG) {
-            Expression::Ptr regAST = makeRegisterExpression(registerTranslation_x86(operand->reg));
-            if (isCFT) {
-                // if a call or a jump has a register as an operand, 
-                // it should not be a conditional jump
-                assert(!isConditional);
-                insn->addSuccessor(regAST, isCall, true, false, false);
-            } else {
-                insn->appendOperand(regAST, 
-                        (operand->access & CS_AC_READ) != 0,
-                        (operand->access & CS_AC_WRITE) != 0, 
-                        false);
-            }
-            //TODO: correctly mark implicit registers
-        } else if (operand->type == X86_OP_IMM) {
-            Expression::Ptr immAST = Immediate::makeImmediate(Result(s32, operand->imm));
-            if (isCFT) {
-                // It looks like that Capstone automatically adjust the offset with the instruction length
-                Expression::Ptr IP(makeRegisterExpression(MachRegister::getPC(m_Arch)));
-                if (isRela) {
-                    Expression::Ptr target(makeAddExpression(IP ,immAST, u64));
-                    insn->addSuccessor(target, isCall, false, isConditional, false);
-                } else {
-                    insn->addSuccessor(immAST, isCall, false, isConditional, false);
-                }
-                if (isConditional)
-                    insn->addSuccessor(IP, false, false, true, true);
-            } else {
-                insn->appendOperand(immAST, false, false, false);
-            }
-        } else if (operand->type == X86_OP_MEM) {
-             Expression::Ptr effectiveAddr;
-             x86_op_mem * mem = &(operand->mem);
-             // TODO: handle segment registers
-             if (mem->base != X86_REG_INVALID) {
-                 effectiveAddr = makeRegisterExpression(registerTranslation_x86(mem->base));
-             }
-             if (mem->index != X86_REG_INVALID) {
-                 Expression::Ptr indexAST = makeRegisterExpression(registerTranslation_x86(mem->index));
-                 indexAST = makeMultiplyExpression(indexAST,
-                             Immediate::makeImmediate(Result(u8, mem->scale)), 
-                             u64);
-                 if (effectiveAddr)
-                     effectiveAddr = makeAddExpression(effectiveAddr, indexAST, u64);
-                 else
-                     effectiveAddr = indexAST;
-             }
-             // Displacement for addressing memory. So it is unsigned
-             Expression::Ptr immAST = Immediate::makeImmediate(Result(u32, mem->disp));
-             if (effectiveAddr)
-                 effectiveAddr = makeAddExpression(effectiveAddr, immAST , u64);
-             else
-                 effectiveAddr = immAST;
-             Result_Type type = operandSizeTranslation(operand->size);
-             if (type == invalid_type) {
-                 err = true;
-             }
-             Expression::Ptr memAST;
-             if (insn->getOperation().getID() == e_lea)
-                 memAST = effectiveAddr;
-             else
-                 memAST = makeDereferenceExpression(effectiveAddr, type);
-             if (isCFT) {
-                 assert(!isConditional);
-                 insn->addSuccessor(memAST, isCall, true, false, false);
-             } else {
-                 insn->appendOperand(memAST, 
-                         (operand->access & CS_AC_READ) != 0,
-                         (operand->access & CS_AC_WRITE) != 0, 
-                         false);
-             }
-        } else {
-            fprintf(stderr, "Unhandled capstone operand type %d\n", operand->type);
-        }
-    }
-    if (err) fprintf(stderr, "\tinstruction %s\n", insn->format().c_str()); 
-
-}
-
-entryID InstructionDecoder_Capstone::opcodeTranslation_x86(unsigned int cap_id) {
-    switch (cap_id) {
-        case X86_INS_AAA:
-            return e_aaa;
-        case X86_INS_AAD:
-            return e_aad;
-        case X86_INS_AAM:
-            return e_aam;
-        case X86_INS_AAS:
-            return e_aas;
-        case X86_INS_FABS:
-            return e_fabs;
-        case X86_INS_ADC:
-            return e_adc;
-        case X86_INS_ADCX:
-            return e_adcx;
-        case X86_INS_ADD:
-            return e_add;
-        case X86_INS_ADDPD:
-            return e_addpd;
-        case X86_INS_ADDPS:
-            return e_addps;
-        case X86_INS_ADDSD:
-            return e_addsd;
-        case X86_INS_ADDSS:
-            return e_addss;
-        case X86_INS_ADDSUBPD:
-            return e_addsubpd;
-        case X86_INS_ADDSUBPS:
-            return e_addsubps;
-        case X86_INS_FADD:
-            return e_fadd;
-        case X86_INS_FIADD:
-            return e_fiadd;
-        case X86_INS_ADOX:
-            return e_adox;
-        case X86_INS_AESDECLAST:
-            return e_aesdeclast;
-        case X86_INS_AESDEC:
-            return e_aesdec;
-        case X86_INS_AESENCLAST:
-            return e_aesenclast;
-        case X86_INS_AESENC:
-            return e_aesenc;
-        case X86_INS_AESIMC:
-            return e_aesimc;
-        case X86_INS_AESKEYGENASSIST:
-            return e_aeskeygenassist;
-        case X86_INS_AND:
-            return e_and;
-        case X86_INS_ANDN:
-            return e_andn;
-        case X86_INS_ANDNPD:
-            return e_andnpd;
-        case X86_INS_ANDNPS:
-            return e_andnps;
-        case X86_INS_ANDPD:
-            return e_andpd;
-        case X86_INS_ANDPS:
-            return e_andps;
-        case X86_INS_ARPL:
-            return e_arpl;
-        case X86_INS_BEXTR:
-            return e_bextr;
-        case X86_INS_BLCFILL:
-            return e_blcfill;
-        case X86_INS_BLCI:
-            return e_blci;
-        case X86_INS_BLCIC:
-            return e_blcic;
-        case X86_INS_BLCMSK:
-            return e_blcmsk;
-        case X86_INS_BLCS:
-            return e_blcs;
-        case X86_INS_BLENDPD:
-            return e_blendpd;
-        case X86_INS_BLENDPS:
-            return e_blendps;
-        case X86_INS_BLENDVPD:
-            return e_blendvpd;
-        case X86_INS_BLENDVPS:
-            return e_blendvps;
-        case X86_INS_BLSFILL:
-            return e_blsfill;
-        case X86_INS_BLSI:
-            return e_blsi;
-        case X86_INS_BLSIC:
-            return e_blsic;
-        case X86_INS_BLSMSK:
-            return e_blsmsk;
-        case X86_INS_BLSR:
-            return e_blsr;
-        case X86_INS_BNDCL:
-            return e_bndcl;
-        case X86_INS_BNDCN:
-            return e_bndcn;
-        case X86_INS_BNDCU:
-            return e_bndcu;
-        case X86_INS_BNDLDX:
-            return e_bndldx;
-        case X86_INS_BNDMK:
-            return e_bndmk;
-        case X86_INS_BNDMOV:
-            return e_bndmov;
-        case X86_INS_BNDSTX:
-            return e_bndstx;
-        case X86_INS_BOUND:
-            return e_bound;
-        case X86_INS_BSF:
-            return e_bsf;
-        case X86_INS_BSR:
-            return e_bsr;
-        case X86_INS_BSWAP:
-            return e_bswap;
-        case X86_INS_BT:
-            return e_bt;
-        case X86_INS_BTC:
-            return e_btc;
-        case X86_INS_BTR:
-            return e_btr;
-        case X86_INS_BTS:
-            return e_bts;
-        case X86_INS_BZHI:
-            return e_bzhi;
-        case X86_INS_CALL:
-            return e_call;
-        case X86_INS_CBW:
-            return e_cbw;
-        case X86_INS_CDQ:
-            return e_cdq;
-        case X86_INS_CDQE:
-            return e_cdqe;
-        case X86_INS_FCHS:
-            return e_fchs;
-        case X86_INS_CLAC:
-            return e_clac;
-        case X86_INS_CLC:
-            return e_clc;
-        case X86_INS_CLD:
-            return e_cld;
-        case X86_INS_CLDEMOTE:
-            return e_cldemote;
-        case X86_INS_CLFLUSH:
-            return e_clflush;
-        case X86_INS_CLFLUSHOPT:
-            return e_clflushopt;
-        case X86_INS_CLGI:
-            return e_clgi;
-        case X86_INS_CLI:
-            return e_cli;
-        case X86_INS_CLRSSBSY:
-            return e_clrssbsy;
-        case X86_INS_CLTS:
-            return e_clts;
-        case X86_INS_CLWB:
-            return e_clwb;
-        case X86_INS_CLZERO:
-            return e_clzero;
-        case X86_INS_CMC:
-            return e_cmc;
-        case X86_INS_CMOVA:
-            return e_cmova;
-        case X86_INS_CMOVAE:
-            return e_cmovae;
-        case X86_INS_CMOVB:
-            return e_cmovb;
-        case X86_INS_CMOVBE:
-            return e_cmovbe;
-        case X86_INS_FCMOVBE:
-            return e_fcmovbe;
-        case X86_INS_FCMOVB:
-            return e_fcmovb;
-        case X86_INS_CMOVE:
-            return e_cmove;
-        case X86_INS_FCMOVE:
-            return e_fcmove;
-        case X86_INS_CMOVG:
-            return e_cmovg;
-        case X86_INS_CMOVGE:
-            return e_cmovge;
-        case X86_INS_CMOVL:
-            return e_cmovl;
-        case X86_INS_CMOVLE:
-            return e_cmovle;
-        case X86_INS_FCMOVNBE:
-            return e_fcmovnbe;
-        case X86_INS_FCMOVNB:
-            return e_fcmovnb;
-        case X86_INS_CMOVNE:
-            return e_cmovne;
-        case X86_INS_FCMOVNE:
-            return e_fcmovne;
-        case X86_INS_CMOVNO:
-            return e_cmovno;
-        case X86_INS_CMOVNP:
-            return e_cmovnp;
-        case X86_INS_FCMOVNU:
-            return e_fcmovnu;
-        case X86_INS_FCMOVNP:
-            return e_fcmovnp;
-        case X86_INS_CMOVNS:
-            return e_cmovns;
-        case X86_INS_CMOVO:
-            return e_cmovo;
-        case X86_INS_CMOVP:
-            return e_cmovp;
-        case X86_INS_FCMOVU:
-            return e_fcmovu;
-        case X86_INS_CMOVS:
-            return e_cmovs;
-        case X86_INS_CMP:
-            return e_cmp;
-        case X86_INS_CMPPD:
-            return e_cmppd;
-        case X86_INS_CMPPS:
-            return e_cmpps;
-        case X86_INS_CMPSB:
-            return e_cmpsb;
-        case X86_INS_CMPSD:
-            return e_cmpsd;
-        case X86_INS_CMPSQ:
-            return e_cmpsq;
-        case X86_INS_CMPSS:
-            return e_cmpss;
-        case X86_INS_CMPSW:
-            return e_cmpsw;
-        case X86_INS_CMPXCHG16B:
-            return e_cmpxchg16b;
-        case X86_INS_CMPXCHG:
-            return e_cmpxchg;
-        case X86_INS_CMPXCHG8B:
-            return e_cmpxchg8b;
-        case X86_INS_COMISD:
-            return e_comisd;
-        case X86_INS_COMISS:
-            return e_comiss;
-        case X86_INS_FCOMP:
-            return e_fcomp;
-        case X86_INS_FCOMPI:
-            return e_fcompi;
-        case X86_INS_FCOMI:
-            return e_fcomi;
-        case X86_INS_FCOM:
-            return e_fcom;
-        case X86_INS_FCOS:
-            return e_fcos;
-        case X86_INS_CPUID:
-            return e_cpuid;
-        case X86_INS_CQO:
-            return e_cqo;
-        case X86_INS_CRC32:
-            return e_crc32;
-        case X86_INS_CVTDQ2PD:
-            return e_cvtdq2pd;
-        case X86_INS_CVTDQ2PS:
-            return e_cvtdq2ps;
-        case X86_INS_CVTPD2DQ:
-            return e_cvtpd2dq;
-        case X86_INS_CVTPD2PS:
-            return e_cvtpd2ps;
-        case X86_INS_CVTPS2DQ:
-            return e_cvtps2dq;
-        case X86_INS_CVTPS2PD:
-            return e_cvtps2pd;
-        case X86_INS_CVTSD2SI:
-            return e_cvtsd2si;
-        case X86_INS_CVTSD2SS:
-            return e_cvtsd2ss;
-        case X86_INS_CVTSI2SD:
-            return e_cvtsi2sd;
-        case X86_INS_CVTSI2SS:
-            return e_cvtsi2ss;
-        case X86_INS_CVTSS2SD:
-            return e_cvtss2sd;
-        case X86_INS_CVTSS2SI:
-            return e_cvtss2si;
-        case X86_INS_CVTTPD2DQ:
-            return e_cvttpd2dq;
-        case X86_INS_CVTTPS2DQ:
-            return e_cvttps2dq;
-        case X86_INS_CVTTSD2SI:
-            return e_cvttsd2si;
-        case X86_INS_CVTTSS2SI:
-            return e_cvttss2si;
-        case X86_INS_CWD:
-            return e_cwd;
-        case X86_INS_CWDE:
-            return e_cwde;
-        case X86_INS_DAA:
-            return e_daa;
-        case X86_INS_DAS:
-            return e_das;
-        case X86_INS_DATA16:
-            return e_data16;
-        case X86_INS_DEC:
-            return e_dec;
-        case X86_INS_DIV:
-            return e_div;
-        case X86_INS_DIVPD:
-            return e_divpd;
-        case X86_INS_DIVPS:
-            return e_divps;
-        case X86_INS_FDIVR:
-            return e_fdivr;
-        case X86_INS_FIDIVR:
-            return e_fidivr;
-        case X86_INS_FDIVRP:
-            return e_fdivrp;
-        case X86_INS_DIVSD:
-            return e_divsd;
-        case X86_INS_DIVSS:
-            return e_divss;
-        case X86_INS_FDIV:
-            return e_fdiv;
-        case X86_INS_FIDIV:
-            return e_fidiv;
-        case X86_INS_FDIVP:
-            return e_fdivp;
-        case X86_INS_DPPD:
-            return e_dppd;
-        case X86_INS_DPPS:
-            return e_dpps;
-        case X86_INS_ENCLS:
-            return e_encls;
-        case X86_INS_ENCLU:
-            return e_enclu;
-        case X86_INS_ENCLV:
-            return e_enclv;
-        case X86_INS_ENDBR32:
-            return e_endbr32;
-        case X86_INS_ENDBR64:
-            return e_endbr64;
-        case X86_INS_ENTER:
-            return e_enter;
-        case X86_INS_EXTRACTPS:
-            return e_extractps;
-        case X86_INS_EXTRQ:
-            return e_extrq;
-        case X86_INS_F2XM1:
-            return e_f2xm1;
-        case X86_INS_LCALL:
-            return e_lcall;
-        case X86_INS_LJMP:
-            return e_ljmp;
-        case X86_INS_JMP:
-            return e_jmp;
-        case X86_INS_FBLD:
-            return e_fbld;
-        case X86_INS_FBSTP:
-            return e_fbstp;
-        case X86_INS_FCOMPP:
-            return e_fcompp;
-        case X86_INS_FDECSTP:
-            return e_fdecstp;
-        case X86_INS_FDISI8087_NOP:
-            return e_nop;
-        case X86_INS_FEMMS:
-            return e_femms;
-        case X86_INS_FENI8087_NOP:
-            return e_nop;
-        case X86_INS_FFREE:
-            return e_ffree;
-        case X86_INS_FFREEP:
-            return e_ffreep;
-        case X86_INS_FICOM:
-            return e_ficom;
-        case X86_INS_FICOMP:
-            return e_ficomp;
-        case X86_INS_FINCSTP:
-            return e_fincstp;
-        case X86_INS_FLDCW:
-            return e_fldcw;
-        case X86_INS_FLDENV:
-            return e_fldenv;
-        case X86_INS_FLDL2E:
-            return e_fldl2e;
-        case X86_INS_FLDL2T:
-            return e_fldl2t;
-        case X86_INS_FLDLG2:
-            return e_fldlg2;
-        case X86_INS_FLDLN2:
-            return e_fldln2;
-        case X86_INS_FLDPI:
-            return e_fldpi;
-        case X86_INS_FNCLEX:
-            return e_fnclex;
-        case X86_INS_FNINIT:
-            return e_fninit;
-        case X86_INS_FNOP:
-            return e_fnop;
-        case X86_INS_FNSTCW:
-            return e_fnstcw;
-        case X86_INS_FNSTSW:
-            return e_fnstsw;
-        case X86_INS_FPATAN:
-            return e_fpatan;
-        case X86_INS_FSTPNCE:
-            return e_fstpnce;
-        case X86_INS_FPREM:
-            return e_fprem;
-        case X86_INS_FPREM1:
-            return e_fprem1;
-        case X86_INS_FPTAN:
-            return e_fptan;
-        case X86_INS_FRNDINT:
-            return e_frndint;
-        case X86_INS_FRSTOR:
-            return e_frstor;
-        case X86_INS_FNSAVE:
-            return e_fnsave;
-        case X86_INS_FSCALE:
-            return e_fscale;
-        case X86_INS_FSETPM:
-            return e_fsetpm;
-        case X86_INS_FSINCOS:
-            return e_fsincos;
-        case X86_INS_FNSTENV:
-            return e_fnstenv;
-        case X86_INS_FXAM:
-            return e_fxam;
-        case X86_INS_FXRSTOR:
-            return e_fxrstor;
-        case X86_INS_FXRSTOR64:
-            return e_fxrstor64;
-        case X86_INS_FXSAVE:
-            return e_fxsave;
-        case X86_INS_FXSAVE64:
-            return e_fxsave64;
-        case X86_INS_FXTRACT:
-            return e_fxtract;
-        case X86_INS_FYL2X:
-            return e_fyl2x;
-        case X86_INS_FYL2XP1:
-            return e_fyl2xp1;
-        case X86_INS_GETSEC:
-            return e_getsec;
-        case X86_INS_GF2P8AFFINEINVQB:
-            return e_gf2p8affineinvqb;
-        case X86_INS_GF2P8AFFINEQB:
-            return e_gf2p8affineqb;
-        case X86_INS_GF2P8MULB:
-            return e_gf2p8mulb;
-        case X86_INS_HADDPD:
-            return e_haddpd;
-        case X86_INS_HADDPS:
-            return e_haddps;
-        case X86_INS_HLT:
-            return e_hlt;
-        case X86_INS_HSUBPD:
-            return e_hsubpd;
-        case X86_INS_HSUBPS:
-            return e_hsubps;
-        case X86_INS_IDIV:
-            return e_idiv;
-        case X86_INS_FILD:
-            return e_fild;
-        case X86_INS_IMUL:
-            return e_imul;
-        case X86_INS_IN:
-            return e_in;
-        case X86_INS_INC:
-            return e_inc;
-        case X86_INS_INCSSPD:
-            return e_incsspd;
-        case X86_INS_INCSSPQ:
-            return e_incsspq;
-        case X86_INS_INSB:
-            return e_insb;
-        case X86_INS_INSERTPS:
-            return e_insertps;
-        case X86_INS_INSERTQ:
-            return e_insertq;
-        case X86_INS_INSD:
-            return e_insd;
-        case X86_INS_INSW:
-            return e_insw;
-        case X86_INS_INT:
-            return e_int;
-        case X86_INS_INT1:
-            return e_int1;
-        case X86_INS_INT3:
-            return e_int3;
-        case X86_INS_INTO:
-            return e_into;
-        case X86_INS_INVD:
-            return e_invd;
-        case X86_INS_INVEPT:
-            return e_invept;
-        case X86_INS_INVLPG:
-            return e_invlpg;
-        case X86_INS_INVLPGA:
-            return e_invlpga;
-        case X86_INS_INVPCID:
-            return e_invpcid;
-        case X86_INS_INVVPID:
-            return e_invvpid;
-        case X86_INS_IRET:
-            return e_iret;
-        case X86_INS_IRETD:
-            return e_iretd;
-        case X86_INS_IRETQ:
-            return e_iretq;
-        case X86_INS_FISTTP:
-            return e_fisttp;
-        case X86_INS_FIST:
-            return e_fist;
-        case X86_INS_FISTP:
-            return e_fistp;
-        case X86_INS_JAE:
-            return e_jnb;
-        case X86_INS_JA:
-            return e_jnbe;
-        case X86_INS_JBE:
-            return e_jbe;
-        case X86_INS_JB:
-            return e_jb;
-        case X86_INS_JCXZ:
-            return e_jcxz;
-        case X86_INS_JECXZ:
-            return e_jecxz;
-        case X86_INS_JE:
-            return e_jz;
-        case X86_INS_JGE:
-            return e_jnl;
-        case X86_INS_JG:
-            return e_jnle;
-        case X86_INS_JLE:
-            return e_jle;
-        case X86_INS_JL:
-            return e_jl;
-        case X86_INS_JNE:
-            return e_jnz;
-        case X86_INS_JNO:
-            return e_jno;
-        case X86_INS_JNP:
-            return e_jnp;
-        case X86_INS_JNS:
-            return e_jns;
-        case X86_INS_JO:
-            return e_jo;
-        case X86_INS_JP:
-            return e_jp;
-        case X86_INS_JRCXZ:
-            return e_jrcxz;
-        case X86_INS_JS:
-            return e_js;
-        case X86_INS_KADDB:
-            return e_kaddb;
-        case X86_INS_KADDD:
-            return e_kaddd;
-        case X86_INS_KADDQ:
-            return e_kaddq;
-        case X86_INS_KADDW:
-            return e_kaddw;
-        case X86_INS_KANDB:
-            return e_kandb;
-        case X86_INS_KANDD:
-            return e_kandd;
-        case X86_INS_KANDNB:
-            return e_kandnb;
-        case X86_INS_KANDND:
-            return e_kandnd;
-        case X86_INS_KANDNQ:
-            return e_kandnq;
-        case X86_INS_KANDNW:
-            return e_kandnw;
-        case X86_INS_KANDQ:
-            return e_kandq;
-        case X86_INS_KANDW:
-            return e_kandw;
-        case X86_INS_KMOVB:
-            return e_kmovb;
-        case X86_INS_KMOVD:
-            return e_kmovd;
-        case X86_INS_KMOVQ:
-            return e_kmovq;
-        case X86_INS_KMOVW:
-            return e_kmovw;
-        case X86_INS_KNOTB:
-            return e_knotb;
-        case X86_INS_KNOTD:
-            return e_knotd;
-        case X86_INS_KNOTQ:
-            return e_knotq;
-        case X86_INS_KNOTW:
-            return e_knotw;
-        case X86_INS_KORB:
-            return e_korb;
-        case X86_INS_KORD:
-            return e_kord;
-        case X86_INS_KORQ:
-            return e_korq;
-        case X86_INS_KORTESTB:
-            return e_kortestb;
-        case X86_INS_KORTESTD:
-            return e_kortestd;
-        case X86_INS_KORTESTQ:
-            return e_kortestq;
-        case X86_INS_KORTESTW:
-            return e_kortestw;
-        case X86_INS_KORW:
-            return e_korw;
-        case X86_INS_KSHIFTLB:
-            return e_kshiftlb;
-        case X86_INS_KSHIFTLD:
-            return e_kshiftld;
-        case X86_INS_KSHIFTLQ:
-            return e_kshiftlq;
-        case X86_INS_KSHIFTLW:
-            return e_kshiftlw;
-        case X86_INS_KSHIFTRB:
-            return e_kshiftrb;
-        case X86_INS_KSHIFTRD:
-            return e_kshiftrd;
-        case X86_INS_KSHIFTRQ:
-            return e_kshiftrq;
-        case X86_INS_KSHIFTRW:
-            return e_kshiftrw;
-        case X86_INS_KTESTB:
-            return e_ktestb;
-        case X86_INS_KTESTD:
-            return e_ktestd;
-        case X86_INS_KTESTQ:
-            return e_ktestq;
-        case X86_INS_KTESTW:
-            return e_ktestw;
-        case X86_INS_KUNPCKBW:
-            return e_kunpckbw;
-        case X86_INS_KUNPCKDQ:
-            return e_kunpckdq;
-        case X86_INS_KUNPCKWD:
-            return e_kunpckwd;
-        case X86_INS_KXNORB:
-            return e_kxnorb;
-        case X86_INS_KXNORD:
-            return e_kxnord;
-        case X86_INS_KXNORQ:
-            return e_kxnorq;
-        case X86_INS_KXNORW:
-            return e_kxnorw;
-        case X86_INS_KXORB:
-            return e_kxorb;
-        case X86_INS_KXORD:
-            return e_kxord;
-        case X86_INS_KXORQ:
-            return e_kxorq;
-        case X86_INS_KXORW:
-            return e_kxorw;
-        case X86_INS_LAHF:
-            return e_lahf;
-        case X86_INS_LAR:
-            return e_lar;
-        case X86_INS_LDDQU:
-            return e_lddqu;
-        case X86_INS_LDMXCSR:
-            return e_ldmxcsr;
-        case X86_INS_LDS:
-            return e_lds;
-        case X86_INS_FLDZ:
-            return e_fldz;
-        case X86_INS_FLD1:
-            return e_fld1;
-        case X86_INS_FLD:
-            return e_fld;
-        case X86_INS_LEA:
-            return e_lea;
-        case X86_INS_LEAVE:
-            return e_leave;
-        case X86_INS_LES:
-            return e_les;
-        case X86_INS_LFENCE:
-            return e_lfence;
-        case X86_INS_LFS:
-            return e_lfs;
-        case X86_INS_LGDT:
-            return e_lgdt;
-        case X86_INS_LGS:
-            return e_lgs;
-        case X86_INS_LIDT:
-            return e_lidt;
-        case X86_INS_LLDT:
-            return e_lldt;
-        case X86_INS_LLWPCB:
-            return e_llwpcb;
-        case X86_INS_LMSW:
-            return e_lmsw;
-        case X86_INS_LOCK:
-            return e_lock;
-        case X86_INS_LODSB:
-            return e_lodsb;
-        case X86_INS_LODSD:
-            return e_lodsd;
-        case X86_INS_LODSQ:
-            return e_lodsq;
-        case X86_INS_LODSW:
-            return e_lodsw;
-        case X86_INS_LOOP:
-            return e_loop;
-        case X86_INS_LOOPE:
-            return e_loope;
-        case X86_INS_LOOPNE:
-            return e_loopne;
-        case X86_INS_RETF:
-            return e_retf;
-        case X86_INS_RETFQ:
-            return e_retfq;
-        case X86_INS_LSL:
-            return e_lsl;
-        case X86_INS_LSS:
-            return e_lss;
-        case X86_INS_LTR:
-            return e_ltr;
-        case X86_INS_LWPINS:
-            return e_lwpins;
-        case X86_INS_LWPVAL:
-            return e_lwpval;
-        case X86_INS_LZCNT:
-            return e_lzcnt;
-        case X86_INS_MASKMOVDQU:
-            return e_maskmovdqu;
-        case X86_INS_MAXPD:
-            return e_maxpd;
-        case X86_INS_MAXPS:
-            return e_maxps;
-        case X86_INS_MAXSD:
-            return e_maxsd;
-        case X86_INS_MAXSS:
-            return e_maxss;
-        case X86_INS_MFENCE:
-            return e_mfence;
-        case X86_INS_MINPD:
-            return e_minpd;
-        case X86_INS_MINPS:
-            return e_minps;
-        case X86_INS_MINSD:
-            return e_minsd;
-        case X86_INS_MINSS:
-            return e_minss;
-        case X86_INS_CVTPD2PI:
-            return e_cvtpd2pi;
-        case X86_INS_CVTPI2PD:
-            return e_cvtpi2pd;
-        case X86_INS_CVTPI2PS:
-            return e_cvtpi2ps;
-        case X86_INS_CVTPS2PI:
-            return e_cvtps2pi;
-        case X86_INS_CVTTPD2PI:
-            return e_cvttpd2pi;
-        case X86_INS_CVTTPS2PI:
-            return e_cvttps2pi;
-        case X86_INS_EMMS:
-            return e_emms;
-        case X86_INS_MASKMOVQ:
-            return e_maskmovq;
-        case X86_INS_MOVD:
-            return e_movd;
-        case X86_INS_MOVQ:
-            return e_movq;
-        case X86_INS_MOVDQ2Q:
-            return e_movdq2q;
-        case X86_INS_MOVNTQ:
-            return e_movntq;
-        case X86_INS_MOVQ2DQ:
-            return e_movq2dq;
-        case X86_INS_PABSB:
-            return e_pabsb;
-        case X86_INS_PABSD:
-            return e_pabsd;
-        case X86_INS_PABSW:
-            return e_pabsw;
-        case X86_INS_PACKSSDW:
-            return e_packssdw;
-        case X86_INS_PACKSSWB:
-            return e_packsswb;
-        case X86_INS_PACKUSWB:
-            return e_packuswb;
-        case X86_INS_PADDB:
-            return e_paddb;
-        case X86_INS_PADDD:
-            return e_paddd;
-        case X86_INS_PADDQ:
-            return e_paddq;
-        case X86_INS_PADDSB:
-            return e_paddsb;
-        case X86_INS_PADDSW:
-            return e_paddsw;
-        case X86_INS_PADDUSB:
-            return e_paddusb;
-        case X86_INS_PADDUSW:
-            return e_paddusw;
-        case X86_INS_PADDW:
-            return e_paddw;
-        case X86_INS_PALIGNR:
-            return e_palignr;
-        case X86_INS_PANDN:
-            return e_pandn;
-        case X86_INS_PAND:
-            return e_pand;
-        case X86_INS_PAVGB:
-            return e_pavgb;
-        case X86_INS_PAVGW:
-            return e_pavgw;
-        case X86_INS_PCMPEQB:
-            return e_pcmpeqb;
-        case X86_INS_PCMPEQD:
-            return e_pcmpeqd;
-        case X86_INS_PCMPEQW:
-            return e_pcmpeqw;
-        case X86_INS_PCMPGTB:
-            return e_pcmpgtb;
-        case X86_INS_PCMPGTD:
-            return e_pcmpgtd;
-        case X86_INS_PCMPGTW:
-            return e_pcmpgtw;
-        case X86_INS_PEXTRW:
-            return e_pextrw;
-        case X86_INS_PHADDD:
-            return e_phaddd;
-        case X86_INS_PHADDSW:
-            return e_phaddsw;
-        case X86_INS_PHADDW:
-            return e_phaddw;
-        case X86_INS_PHSUBD:
-            return e_phsubd;
-        case X86_INS_PHSUBSW:
-            return e_phsubsw;
-        case X86_INS_PHSUBW:
-            return e_phsubw;
-        case X86_INS_PINSRW:
-            return e_pinsrw;
-        case X86_INS_PMADDUBSW:
-            return e_pmaddubsw;
-        case X86_INS_PMADDWD:
-            return e_pmaddwd;
-        case X86_INS_PMAXSW:
-            return e_pmaxsw;
-        case X86_INS_PMAXUB:
-            return e_pmaxub;
-        case X86_INS_PMINSW:
-            return e_pminsw;
-        case X86_INS_PMINUB:
-            return e_pminub;
-        case X86_INS_PMOVMSKB:
-            return e_pmovmskb;
-        case X86_INS_PMULHRSW:
-            return e_pmulhrsw;
-        case X86_INS_PMULHUW:
-            return e_pmulhuw;
-        case X86_INS_PMULHW:
-            return e_pmulhw;
-        case X86_INS_PMULLW:
-            return e_pmullw;
-        case X86_INS_PMULUDQ:
-            return e_pmuludq;
-        case X86_INS_POR:
-            return e_por;
-        case X86_INS_PSADBW:
-            return e_psadbw;
-        case X86_INS_PSHUFB:
-            return e_pshufb;
-        case X86_INS_PSHUFW:
-            return e_pshufw;
-        case X86_INS_PSIGNB:
-            return e_psignb;
-        case X86_INS_PSIGND:
-            return e_psignd;
-        case X86_INS_PSIGNW:
-            return e_psignw;
-        case X86_INS_PSLLD:
-            return e_pslld;
-        case X86_INS_PSLLQ:
-            return e_psllq;
-        case X86_INS_PSLLW:
-            return e_psllw;
-        case X86_INS_PSRAD:
-            return e_psrad;
-        case X86_INS_PSRAW:
-            return e_psraw;
-        case X86_INS_PSRLD:
-            return e_psrld;
-        case X86_INS_PSRLQ:
-            return e_psrlq;
-        case X86_INS_PSRLW:
-            return e_psrlw;
-        case X86_INS_PSUBB:
-            return e_psubb;
-        case X86_INS_PSUBD:
-            return e_psubd;
-        case X86_INS_PSUBQ:
-            return e_psubq;
-        case X86_INS_PSUBSB:
-            return e_psubsb;
-        case X86_INS_PSUBSW:
-            return e_psubsw;
-        case X86_INS_PSUBUSB:
-            return e_psubusb;
-        case X86_INS_PSUBUSW:
-            return e_psubusw;
-        case X86_INS_PSUBW:
-            return e_psubw;
-        case X86_INS_PUNPCKHBW:
-            return e_punpckhbw;
-        case X86_INS_PUNPCKHDQ:
-            return e_punpckhdq;
-        case X86_INS_PUNPCKHWD:
-            return e_punpckhwd;
-        case X86_INS_PUNPCKLBW:
-            return e_punpcklbw;
-        case X86_INS_PUNPCKLDQ:
-            return e_punpckldq;
-        case X86_INS_PUNPCKLWD:
-            return e_punpcklwd;
-        case X86_INS_PXOR:
-            return e_pxor;
-        case X86_INS_MONITORX:
-            return e_monitorx;
-        case X86_INS_MONITOR:
-            return e_monitor;
-        case X86_INS_MONTMUL:
-            return e_montmul;
-        case X86_INS_MOV:
-            return e_mov;
-        case X86_INS_MOVABS:
-            return e_movabs;
-        case X86_INS_MOVAPD:
-            return e_movapd;
-        case X86_INS_MOVAPS:
-            return e_movaps;
-        case X86_INS_MOVBE:
-            return e_movbe;
-        case X86_INS_MOVDDUP:
-            return e_movddup;
-        case X86_INS_MOVDIR64B:
-            return e_movdir64b;
-        case X86_INS_MOVDIRI:
-            return e_movdiri;
-        case X86_INS_MOVDQA:
-            return e_movdqa;
-        case X86_INS_MOVDQU:
-            return e_movdqu;
-        case X86_INS_MOVHLPS:
-            return e_movhlps;
-        case X86_INS_MOVHPD:
-            return e_movhpd;
-        case X86_INS_MOVHPS:
-            return e_movhps;
-        case X86_INS_MOVLHPS:
-            return e_movlhps;
-        case X86_INS_MOVLPD:
-            return e_movlpd;
-        case X86_INS_MOVLPS:
-            return e_movlps;
-        case X86_INS_MOVMSKPD:
-            return e_movmskpd;
-        case X86_INS_MOVMSKPS:
-            return e_movmskps;
-        case X86_INS_MOVNTDQA:
-            return e_movntdqa;
-        case X86_INS_MOVNTDQ:
-            return e_movntdq;
-        case X86_INS_MOVNTI:
-            return e_movnti;
-        case X86_INS_MOVNTPD:
-            return e_movntpd;
-        case X86_INS_MOVNTPS:
-            return e_movntps;
-        case X86_INS_MOVNTSD:
-            return e_movntsd;
-        case X86_INS_MOVNTSS:
-            return e_movntss;
-        case X86_INS_MOVSB:
-            return e_movsb;
-        case X86_INS_MOVSD:
-            return e_movsd;
-        case X86_INS_MOVSHDUP:
-            return e_movshdup;
-        case X86_INS_MOVSLDUP:
-            return e_movsldup;
-        case X86_INS_MOVSQ:
-            return e_movsq;
-        case X86_INS_MOVSS:
-            return e_movss;
-        case X86_INS_MOVSW:
-            return e_movsw;
-        case X86_INS_MOVSX:
-            return e_movsx;
-        case X86_INS_MOVSXD:
-            return e_movsxd;
-        case X86_INS_MOVUPD:
-            return e_movupd;
-        case X86_INS_MOVUPS:
-            return e_movups;
-        case X86_INS_MOVZX:
-            return e_movzx;
-        case X86_INS_MPSADBW:
-            return e_mpsadbw;
-        case X86_INS_MUL:
-            return e_mul;
-        case X86_INS_MULPD:
-            return e_mulpd;
-        case X86_INS_MULPS:
-            return e_mulps;
-        case X86_INS_MULSD:
-            return e_mulsd;
-        case X86_INS_MULSS:
-            return e_mulss;
-        case X86_INS_MULX:
-            return e_mulx;
-        case X86_INS_FMUL:
-            return e_fmul;
-        case X86_INS_FIMUL:
-            return e_fimul;
-        case X86_INS_FMULP:
-            return e_fmulp;
-        case X86_INS_MWAITX:
-            return e_mwaitx;
-        case X86_INS_MWAIT:
-            return e_mwait;
-        case X86_INS_NEG:
-            return e_neg;
-        case X86_INS_NOP:
-            return e_nop;
-        case X86_INS_NOT:
-            return e_not;
-        case X86_INS_OR:
-            return e_or;
-        case X86_INS_ORPD:
-            return e_orpd;
-        case X86_INS_ORPS:
-            return e_orps;
-        case X86_INS_OUT:
-            return e_out;
-        case X86_INS_OUTSB:
-            return e_outsb;
-        case X86_INS_OUTSD:
-            return e_outsd;
-        case X86_INS_OUTSW:
-            return e_outsw;
-        case X86_INS_PACKUSDW:
-            return e_packusdw;
-        case X86_INS_PAUSE:
-            return e_pause;
-        case X86_INS_PAVGUSB:
-            return e_pavgusb;
-        case X86_INS_PBLENDVB:
-            return e_pblendvb;
-        case X86_INS_PBLENDW:
-            return e_pblendw;
-        case X86_INS_PCLMULQDQ:
-            return e_pclmulqdq;
-        case X86_INS_PCMPEQQ:
-            return e_pcmpeqq;
-        case X86_INS_PCMPESTRI:
-            return e_pcmpestri;
-        case X86_INS_PCMPESTRM:
-            return e_pcmpestrm;
-        case X86_INS_PCMPGTQ:
-            return e_pcmpgtq;
-        case X86_INS_PCMPISTRI:
-            return e_pcmpistri;
-        case X86_INS_PCMPISTRM:
-            return e_pcmpistrm;
-        case X86_INS_PCONFIG:
-            return e_pconfig;
-        case X86_INS_PDEP:
-            return e_pdep;
-        case X86_INS_PEXT:
-            return e_pext;
-        case X86_INS_PEXTRB:
-            return e_pextrb;
-        case X86_INS_PEXTRD:
-            return e_pextrd;
-        case X86_INS_PEXTRQ:
-            return e_pextrq;
-        case X86_INS_PF2ID:
-            return e_pf2id;
-        case X86_INS_PF2IW:
-            return e_pf2iw;
-        case X86_INS_PFACC:
-            return e_pfacc;
-        case X86_INS_PFADD:
-            return e_pfadd;
-        case X86_INS_PFCMPEQ:
-            return e_pfcmpeq;
-        case X86_INS_PFCMPGE:
-            return e_pfcmpge;
-        case X86_INS_PFCMPGT:
-            return e_pfcmpgt;
-        case X86_INS_PFMAX:
-            return e_pfmax;
-        case X86_INS_PFMIN:
-            return e_pfmin;
-        case X86_INS_PFMUL:
-            return e_pfmul;
-        case X86_INS_PFNACC:
-            return e_pfnacc;
-        case X86_INS_PFPNACC:
-            return e_pfpnacc;
-        case X86_INS_PFRCPIT1:
-            return e_pfrcpit1;
-        case X86_INS_PFRCPIT2:
-            return e_pfrcpit2;
-        case X86_INS_PFRCP:
-            return e_pfrcp;
-        case X86_INS_PFRSQIT1:
-            return e_pfrsqit1;
-        case X86_INS_PFRSQRT:
-            return e_pfrsqrt;
-        case X86_INS_PFSUBR:
-            return e_pfsubr;
-        case X86_INS_PFSUB:
-            return e_pfsub;
-        case X86_INS_PHMINPOSUW:
-            return e_phminposuw;
-        case X86_INS_PI2FD:
-            return e_pi2fd;
-        case X86_INS_PI2FW:
-            return e_pi2fw;
-        case X86_INS_PINSRB:
-            return e_pinsrb;
-        case X86_INS_PINSRD:
-            return e_pinsrd;
-        case X86_INS_PINSRQ:
-            return e_pinsrq;
-        case X86_INS_PMAXSB:
-            return e_pmaxsb;
-        case X86_INS_PMAXSD:
-            return e_pmaxsd;
-        case X86_INS_PMAXUD:
-            return e_pmaxud;
-        case X86_INS_PMAXUW:
-            return e_pmaxuw;
-        case X86_INS_PMINSB:
-            return e_pminsb;
-        case X86_INS_PMINSD:
-            return e_pminsd;
-        case X86_INS_PMINUD:
-            return e_pminud;
-        case X86_INS_PMINUW:
-            return e_pminuw;
-        case X86_INS_PMOVSXBD:
-            return e_pmovsxbd;
-        case X86_INS_PMOVSXBQ:
-            return e_pmovsxbq;
-        case X86_INS_PMOVSXBW:
-            return e_pmovsxbw;
-        case X86_INS_PMOVSXDQ:
-            return e_pmovsxdq;
-        case X86_INS_PMOVSXWD:
-            return e_pmovsxwd;
-        case X86_INS_PMOVSXWQ:
-            return e_pmovsxwq;
-        case X86_INS_PMOVZXBD:
-            return e_pmovzxbd;
-        case X86_INS_PMOVZXBQ:
-            return e_pmovzxbq;
-        case X86_INS_PMOVZXBW:
-            return e_pmovzxbw;
-        case X86_INS_PMOVZXDQ:
-            return e_pmovzxdq;
-        case X86_INS_PMOVZXWD:
-            return e_pmovzxwd;
-        case X86_INS_PMOVZXWQ:
-            return e_pmovzxwq;
-        case X86_INS_PMULDQ:
-            return e_pmuldq;
-        case X86_INS_PMULHRW:
-            return e_pmulhrw;
-        case X86_INS_PMULLD:
-            return e_pmulld;
-        case X86_INS_POP:
-            return e_pop;
-        case X86_INS_POPAW:
-            return e_popaw;
-        case X86_INS_POPAL:
-            return e_popal;
-        case X86_INS_POPCNT:
-            return e_popcnt;
-        case X86_INS_POPF:
-            return e_popf;
-        case X86_INS_POPFD:
-            return e_popfd;
-        case X86_INS_POPFQ:
-            return e_popfq;
-        case X86_INS_PREFETCH:
-            return e_prefetch;
-        case X86_INS_PREFETCHNTA:
-            return e_prefetchnta;
-        case X86_INS_PREFETCHT0:
-            return e_prefetcht0;
-        case X86_INS_PREFETCHT1:
-            return e_prefetcht1;
-        case X86_INS_PREFETCHT2:
-            return e_prefetcht2;
-        case X86_INS_PREFETCHW:
-            return e_prefetchw;
-        case X86_INS_PREFETCHWT1:
-            return e_prefetchwt1;
-        case X86_INS_PSHUFD:
-            return e_pshufd;
-        case X86_INS_PSHUFHW:
-            return e_pshufhw;
-        case X86_INS_PSHUFLW:
-            return e_pshuflw;
-        case X86_INS_PSLLDQ:
-            return e_pslldq;
-        case X86_INS_PSRLDQ:
-            return e_psrldq;
-        case X86_INS_PSWAPD:
-            return e_pswapd;
-        case X86_INS_PTEST:
-            return e_ptest;
-        case X86_INS_PTWRITE:
-            return e_ptwrite;
-        case X86_INS_PUNPCKHQDQ:
-            return e_punpckhqdq;
-        case X86_INS_PUNPCKLQDQ:
-            return e_punpcklqdq;
-        case X86_INS_PUSH:
-            return e_push;
-        case X86_INS_PUSHAW:
-            return e_pushaw;
-        case X86_INS_PUSHAL:
-            return e_pushal;
-        case X86_INS_PUSHF:
-            return e_pushf;
-        case X86_INS_PUSHFD:
-            return e_pushfd;
-        case X86_INS_PUSHFQ:
-            return e_pushfq;
-        case X86_INS_RCL:
-            return e_rcl;
-        case X86_INS_RCPPS:
-            return e_rcpps;
-        case X86_INS_RCPSS:
-            return e_rcpss;
-        case X86_INS_RCR:
-            return e_rcr;
-        case X86_INS_RDFSBASE:
-            return e_rdfsbase;
-        case X86_INS_RDGSBASE:
-            return e_rdgsbase;
-        case X86_INS_RDMSR:
-            return e_rdmsr;
-        case X86_INS_RDPID:
-            return e_rdpid;
-        case X86_INS_RDPKRU:
-            return e_rdpkru;
-        case X86_INS_RDPMC:
-            return e_rdpmc;
-        case X86_INS_RDRAND:
-            return e_rdrand;
-        case X86_INS_RDSEED:
-            return e_rdseed;
-        case X86_INS_RDSSPD:
-            return e_rdsspd;
-        case X86_INS_RDSSPQ:
-            return e_rdsspq;
-        case X86_INS_RDTSC:
-            return e_rdtsc;
-        case X86_INS_RDTSCP:
-            return e_rdtscp;
-        case X86_INS_REPNE:
-            return e_repne;
-        case X86_INS_REP:
-            return e_rep;
-        case X86_INS_RET:
-            return e_ret_near;
-        case X86_INS_REX64:
-            return e_rex64;
-        case X86_INS_ROL:
-            return e_rol;
-        case X86_INS_ROR:
-            return e_ror;
-        case X86_INS_RORX:
-            return e_rorx;
-        case X86_INS_ROUNDPD:
-            return e_roundpd;
-        case X86_INS_ROUNDPS:
-            return e_roundps;
-        case X86_INS_ROUNDSD:
-            return e_roundsd;
-        case X86_INS_ROUNDSS:
-            return e_roundss;
-        case X86_INS_RSM:
-            return e_rsm;
-        case X86_INS_RSQRTPS:
-            return e_rsqrtps;
-        case X86_INS_RSQRTSS:
-            return e_rsqrtss;
-        case X86_INS_RSTORSSP:
-            return e_rstorssp;
-        case X86_INS_SAHF:
-            return e_sahf;
-        case X86_INS_SAL:
-            return e_sal;
-        case X86_INS_SALC:
-            return e_salc;
-        case X86_INS_SAR:
-            return e_sar;
-        case X86_INS_SARX:
-            return e_sarx;
-        case X86_INS_SAVEPREVSSP:
-            return e_saveprevssp;
-        case X86_INS_SBB:
-            return e_sbb;
-        case X86_INS_SCASB:
-            return e_scasb;
-        case X86_INS_SCASD:
-            return e_scasd;
-        case X86_INS_SCASQ:
-            return e_scasq;
-        case X86_INS_SCASW:
-            return e_scasw;
-        case X86_INS_SETAE:
-            return e_setae;
-        case X86_INS_SETA:
-            return e_seta;
-        case X86_INS_SETBE:
-            return e_setbe;
-        case X86_INS_SETB:
-            return e_setb;
-        case X86_INS_SETE:
-            return e_sete;
-        case X86_INS_SETGE:
-            return e_setge;
-        case X86_INS_SETG:
-            return e_setg;
-        case X86_INS_SETLE:
-            return e_setle;
-        case X86_INS_SETL:
-            return e_setl;
-        case X86_INS_SETNE:
-            return e_setne;
-        case X86_INS_SETNO:
-            return e_setno;
-        case X86_INS_SETNP:
-            return e_setnp;
-        case X86_INS_SETNS:
-            return e_setns;
-        case X86_INS_SETO:
-            return e_seto;
-        case X86_INS_SETP:
-            return e_setp;
-        case X86_INS_SETSSBSY:
-            return e_setssbsy;
-        case X86_INS_SETS:
-            return e_sets;
-        case X86_INS_SFENCE:
-            return e_sfence;
-        case X86_INS_SGDT:
-            return e_sgdt;
-        case X86_INS_SHA1MSG1:
-            return e_sha1msg1;
-        case X86_INS_SHA1MSG2:
-            return e_sha1msg2;
-        case X86_INS_SHA1NEXTE:
-            return e_sha1nexte;
-        case X86_INS_SHA1RNDS4:
-            return e_sha1rnds4;
-        case X86_INS_SHA256MSG1:
-            return e_sha256msg1;
-        case X86_INS_SHA256MSG2:
-            return e_sha256msg2;
-        case X86_INS_SHA256RNDS2:
-            return e_sha256rnds2;
-        case X86_INS_SHL:
-            return e_shl;
-        case X86_INS_SHLD:
-            return e_shld;
-        case X86_INS_SHLX:
-            return e_shlx;
-        case X86_INS_SHR:
-            return e_shr;
-        case X86_INS_SHRD:
-            return e_shrd;
-        case X86_INS_SHRX:
-            return e_shrx;
-        case X86_INS_SHUFPD:
-            return e_shufpd;
-        case X86_INS_SHUFPS:
-            return e_shufps;
-        case X86_INS_SIDT:
-            return e_sidt;
-        case X86_INS_FSIN:
-            return e_fsin;
-        case X86_INS_SKINIT:
-            return e_skinit;
-        case X86_INS_SLDT:
-            return e_sldt;
-        case X86_INS_SLWPCB:
-            return e_slwpcb;
-        case X86_INS_SMSW:
-            return e_smsw;
-        case X86_INS_SQRTPD:
-            return e_sqrtpd;
-        case X86_INS_SQRTPS:
-            return e_sqrtps;
-        case X86_INS_SQRTSD:
-            return e_sqrtsd;
-        case X86_INS_SQRTSS:
-            return e_sqrtss;
-        case X86_INS_FSQRT:
-            return e_fsqrt;
-        case X86_INS_STAC:
-            return e_stac;
-        case X86_INS_STC:
-            return e_stc;
-        case X86_INS_STD:
-            return e_std;
-        case X86_INS_STGI:
-            return e_stgi;
-        case X86_INS_STI:
-            return e_sti;
-        case X86_INS_STMXCSR:
-            return e_stmxcsr;
-        case X86_INS_STOSB:
-            return e_stosb;
-        case X86_INS_STOSD:
-            return e_stosd;
-        case X86_INS_STOSQ:
-            return e_stosq;
-        case X86_INS_STOSW:
-            return e_stosw;
-        case X86_INS_STR:
-            return e_str;
-        case X86_INS_FST:
-            return e_fst;
-        case X86_INS_FSTP:
-            return e_fstp;
-        case X86_INS_SUB:
-            return e_sub;
-        case X86_INS_SUBPD:
-            return e_subpd;
-        case X86_INS_SUBPS:
-            return e_subps;
-        case X86_INS_FSUBR:
-            return e_fsubr;
-        case X86_INS_FISUBR:
-            return e_fisubr;
-        case X86_INS_FSUBRP:
-            return e_fsubrp;
-        case X86_INS_SUBSD:
-            return e_subsd;
-        case X86_INS_SUBSS:
-            return e_subss;
-        case X86_INS_FSUB:
-            return e_fsub;
-        case X86_INS_FISUB:
-            return e_fisub;
-        case X86_INS_FSUBP:
-            return e_fsubp;
-        case X86_INS_SWAPGS:
-            return e_swapgs;
-        case X86_INS_SYSCALL:
-            return e_syscall;
-        case X86_INS_SYSENTER:
-            return e_sysenter;
-        case X86_INS_SYSEXIT:
-            return e_sysexit;
-        case X86_INS_SYSEXITQ:
-            return e_sysexitq;
-        case X86_INS_SYSRET:
-            return e_sysret;
-        case X86_INS_SYSRETQ:
-            return e_sysretq;
-        case X86_INS_T1MSKC:
-            return e_t1mskc;
-        case X86_INS_TEST:
-            return e_test;
-        case X86_INS_TPAUSE:
-            return e_tpause;
-        case X86_INS_FTST:
-            return e_ftst;
-        case X86_INS_TZCNT:
-            return e_tzcnt;
-        case X86_INS_TZMSK:
-            return e_tzmsk;
-        case X86_INS_UCOMISD:
-            return e_ucomisd;
-        case X86_INS_UCOMISS:
-            return e_ucomiss;
-        case X86_INS_FUCOMPI:
-            return e_fucompi;
-        case X86_INS_FUCOMI:
-            return e_fucomi;
-        case X86_INS_FUCOMPP:
-            return e_fucompp;
-        case X86_INS_FUCOMP:
-            return e_fucomp;
-        case X86_INS_FUCOM:
-            return e_fucom;
-        case X86_INS_UD0:
-            return e_ud0;
-        case X86_INS_UD1:
-            return e_ud1;
-        case X86_INS_UD2:
-            return e_ud2;
-        case X86_INS_UMONITOR:
-            return e_umonitor;
-        case X86_INS_UMWAIT:
-            return e_umwait;
-        case X86_INS_UNPCKHPD:
-            return e_unpckhpd;
-        case X86_INS_UNPCKHPS:
-            return e_unpckhps;
-        case X86_INS_UNPCKLPD:
-            return e_unpcklpd;
-        case X86_INS_UNPCKLPS:
-            return e_unpcklps;
-        case X86_INS_V4FMADDPS:
-            return e_v4fmaddps;
-        case X86_INS_V4FMADDSS:
-            return e_v4fmaddss;
-        case X86_INS_V4FNMADDPS:
-            return e_v4fnmaddps;
-        case X86_INS_V4FNMADDSS:
-            return e_v4fnmaddss;
-        case X86_INS_VADDPD:
-            return e_vaddpd;
-        case X86_INS_VADDPS:
-            return e_vaddps;
-        case X86_INS_VADDSD:
-            return e_vaddsd;
-        case X86_INS_VADDSS:
-            return e_vaddss;
-        case X86_INS_VADDSUBPD:
-            return e_vaddsubpd;
-        case X86_INS_VADDSUBPS:
-            return e_vaddsubps;
-        case X86_INS_VAESDECLAST:
-            return e_vaesdeclast;
-        case X86_INS_VAESDEC:
-            return e_vaesdec;
-        case X86_INS_VAESENCLAST:
-            return e_vaesenclast;
-        case X86_INS_VAESENC:
-            return e_vaesenc;
-        case X86_INS_VAESIMC:
-            return e_vaesimc;
-        case X86_INS_VAESKEYGENASSIST:
-            return e_vaeskeygenassist;
-        case X86_INS_VALIGND:
-            return e_valignd;
-        case X86_INS_VALIGNQ:
-            return e_valignq;
-        case X86_INS_VANDNPD:
-            return e_vandnpd;
-        case X86_INS_VANDNPS:
-            return e_vandnps;
-        case X86_INS_VANDPD:
-            return e_vandpd;
-        case X86_INS_VANDPS:
-            return e_vandps;
-        case X86_INS_VBLENDMPD:
-            return e_vblendmpd;
-        case X86_INS_VBLENDMPS:
-            return e_vblendmps;
-        case X86_INS_VBLENDPD:
-            return e_vblendpd;
-        case X86_INS_VBLENDPS:
-            return e_vblendps;
-        case X86_INS_VBLENDVPD:
-            return e_vblendvpd;
-        case X86_INS_VBLENDVPS:
-            return e_vblendvps;
-        case X86_INS_VBROADCASTF128:
-            return e_vbroadcastf128;
-        case X86_INS_VBROADCASTF32X2:
-            return e_vbroadcastf32x2;
-        case X86_INS_VBROADCASTF32X4:
-            return e_vbroadcastf32x4;
-        case X86_INS_VBROADCASTF32X8:
-            return e_vbroadcastf32x8;
-        case X86_INS_VBROADCASTF64X2:
-            return e_vbroadcastf64x2;
-        case X86_INS_VBROADCASTF64X4:
-            return e_vbroadcastf64x4;
-        case X86_INS_VBROADCASTI128:
-            return e_vbroadcasti128;
-        case X86_INS_VBROADCASTI32X2:
-            return e_vbroadcasti32x2;
-        case X86_INS_VBROADCASTI32X4:
-            return e_vbroadcasti32x4;
-        case X86_INS_VBROADCASTI32X8:
-            return e_vbroadcasti32x8;
-        case X86_INS_VBROADCASTI64X2:
-            return e_vbroadcasti64x2;
-        case X86_INS_VBROADCASTI64X4:
-            return e_vbroadcasti64x4;
-        case X86_INS_VBROADCASTSD:
-            return e_vbroadcastsd;
-        case X86_INS_VBROADCASTSS:
-            return e_vbroadcastss;
-        case X86_INS_VCMP:
-            return e_vcmp;
-        case X86_INS_VCMPPD:
-            return e_vcmppd;
-        case X86_INS_VCMPPS:
-            return e_vcmpps;
-        case X86_INS_VCMPSD:
-            return e_vcmpsd;
-        case X86_INS_VCMPSS:
-            return e_vcmpss;
-        case X86_INS_VCOMISD:
-            return e_vcomisd;
-        case X86_INS_VCOMISS:
-            return e_vcomiss;
-        case X86_INS_VCOMPRESSPD:
-            return e_vcompresspd;
-        case X86_INS_VCOMPRESSPS:
-            return e_vcompressps;
-        case X86_INS_VCVTDQ2PD:
-            return e_vcvtdq2pd;
-        case X86_INS_VCVTDQ2PS:
-            return e_vcvtdq2ps;
-        case X86_INS_VCVTPD2DQ:
-            return e_vcvtpd2dq;
-        case X86_INS_VCVTPD2PS:
-            return e_vcvtpd2ps;
-        case X86_INS_VCVTPD2QQ:
-            return e_vcvtpd2qq;
-        case X86_INS_VCVTPD2UDQ:
-            return e_vcvtpd2udq;
-        case X86_INS_VCVTPD2UQQ:
-            return e_vcvtpd2uqq;
-        case X86_INS_VCVTPH2PS:
-            return e_vcvtph2ps;
-        case X86_INS_VCVTPS2DQ:
-            return e_vcvtps2dq;
-        case X86_INS_VCVTPS2PD:
-            return e_vcvtps2pd;
-        case X86_INS_VCVTPS2PH:
-            return e_vcvtps2ph;
-        case X86_INS_VCVTPS2QQ:
-            return e_vcvtps2qq;
-        case X86_INS_VCVTPS2UDQ:
-            return e_vcvtps2udq;
-        case X86_INS_VCVTPS2UQQ:
-            return e_vcvtps2uqq;
-        case X86_INS_VCVTQQ2PD:
-            return e_vcvtqq2pd;
-        case X86_INS_VCVTQQ2PS:
-            return e_vcvtqq2ps;
-        case X86_INS_VCVTSD2SI:
-            return e_vcvtsd2si;
-        case X86_INS_VCVTSD2SS:
-            return e_vcvtsd2ss;
-        case X86_INS_VCVTSD2USI:
-            return e_vcvtsd2usi;
-        case X86_INS_VCVTSI2SD:
-            return e_vcvtsi2sd;
-        case X86_INS_VCVTSI2SS:
-            return e_vcvtsi2ss;
-        case X86_INS_VCVTSS2SD:
-            return e_vcvtss2sd;
-        case X86_INS_VCVTSS2SI:
-            return e_vcvtss2si;
-        case X86_INS_VCVTSS2USI:
-            return e_vcvtss2usi;
-        case X86_INS_VCVTTPD2DQ:
-            return e_vcvttpd2dq;
-        case X86_INS_VCVTTPD2QQ:
-            return e_vcvttpd2qq;
-        case X86_INS_VCVTTPD2UDQ:
-            return e_vcvttpd2udq;
-        case X86_INS_VCVTTPD2UQQ:
-            return e_vcvttpd2uqq;
-        case X86_INS_VCVTTPS2DQ:
-            return e_vcvttps2dq;
-        case X86_INS_VCVTTPS2QQ:
-            return e_vcvttps2qq;
-        case X86_INS_VCVTTPS2UDQ:
-            return e_vcvttps2udq;
-        case X86_INS_VCVTTPS2UQQ:
-            return e_vcvttps2uqq;
-        case X86_INS_VCVTTSD2SI:
-            return e_vcvttsd2si;
-        case X86_INS_VCVTTSD2USI:
-            return e_vcvttsd2usi;
-        case X86_INS_VCVTTSS2SI:
-            return e_vcvttss2si;
-        case X86_INS_VCVTTSS2USI:
-            return e_vcvttss2usi;
-        case X86_INS_VCVTUDQ2PD:
-            return e_vcvtudq2pd;
-        case X86_INS_VCVTUDQ2PS:
-            return e_vcvtudq2ps;
-        case X86_INS_VCVTUQQ2PD:
-            return e_vcvtuqq2pd;
-        case X86_INS_VCVTUQQ2PS:
-            return e_vcvtuqq2ps;
-        case X86_INS_VCVTUSI2SD:
-            return e_vcvtusi2sd;
-        case X86_INS_VCVTUSI2SS:
-            return e_vcvtusi2ss;
-        case X86_INS_VDBPSADBW:
-            return e_vdbpsadbw;
-        case X86_INS_VDIVPD:
-            return e_vdivpd;
-        case X86_INS_VDIVPS:
-            return e_vdivps;
-        case X86_INS_VDIVSD:
-            return e_vdivsd;
-        case X86_INS_VDIVSS:
-            return e_vdivss;
-        case X86_INS_VDPPD:
-            return e_vdppd;
-        case X86_INS_VDPPS:
-            return e_vdpps;
-        case X86_INS_VERR:
-            return e_verr;
-        case X86_INS_VERW:
-            return e_verw;
-        case X86_INS_VEXP2PD:
-            return e_vexp2pd;
-        case X86_INS_VEXP2PS:
-            return e_vexp2ps;
-        case X86_INS_VEXPANDPD:
-            return e_vexpandpd;
-        case X86_INS_VEXPANDPS:
-            return e_vexpandps;
-        case X86_INS_VEXTRACTF128:
-            return e_vextractf128;
-        case X86_INS_VEXTRACTF32X4:
-            return e_vextractf32x4;
-        case X86_INS_VEXTRACTF32X8:
-            return e_vextractf32x8;
-        case X86_INS_VEXTRACTF64X2:
-            return e_vextractf64x2;
-        case X86_INS_VEXTRACTF64X4:
-            return e_vextractf64x4;
-        case X86_INS_VEXTRACTI128:
-            return e_vextracti128;
-        case X86_INS_VEXTRACTI32X4:
-            return e_vextracti32x4;
-        case X86_INS_VEXTRACTI32X8:
-            return e_vextracti32x8;
-        case X86_INS_VEXTRACTI64X2:
-            return e_vextracti64x2;
-        case X86_INS_VEXTRACTI64X4:
-            return e_vextracti64x4;
-        case X86_INS_VEXTRACTPS:
-            return e_vextractps;
-        case X86_INS_VFIXUPIMMPD:
-            return e_vfixupimmpd;
-        case X86_INS_VFIXUPIMMPS:
-            return e_vfixupimmps;
-        case X86_INS_VFIXUPIMMSD:
-            return e_vfixupimmsd;
-        case X86_INS_VFIXUPIMMSS:
-            return e_vfixupimmss;
-        case X86_INS_VFMADD132PD:
-            return e_vfmadd132pd;
-        case X86_INS_VFMADD132PS:
-            return e_vfmadd132ps;
-        case X86_INS_VFMADD132SD:
-            return e_vfmadd132sd;
-        case X86_INS_VFMADD132SS:
-            return e_vfmadd132ss;
-        case X86_INS_VFMADD213PD:
-            return e_vfmadd213pd;
-        case X86_INS_VFMADD213PS:
-            return e_vfmadd213ps;
-        case X86_INS_VFMADD213SD:
-            return e_vfmadd213sd;
-        case X86_INS_VFMADD213SS:
-            return e_vfmadd213ss;
-        case X86_INS_VFMADD231PD:
-            return e_vfmadd231pd;
-        case X86_INS_VFMADD231PS:
-            return e_vfmadd231ps;
-        case X86_INS_VFMADD231SD:
-            return e_vfmadd231sd;
-        case X86_INS_VFMADD231SS:
-            return e_vfmadd231ss;
-        case X86_INS_VFMADDPD:
-            return e_vfmaddpd;
-        case X86_INS_VFMADDPS:
-            return e_vfmaddps;
-        case X86_INS_VFMADDSD:
-            return e_vfmaddsd;
-        case X86_INS_VFMADDSS:
-            return e_vfmaddss;
-        case X86_INS_VFMADDSUB132PD:
-            return e_vfmaddsub132pd;
-        case X86_INS_VFMADDSUB132PS:
-            return e_vfmaddsub132ps;
-        case X86_INS_VFMADDSUB213PD:
-            return e_vfmaddsub213pd;
-        case X86_INS_VFMADDSUB213PS:
-            return e_vfmaddsub213ps;
-        case X86_INS_VFMADDSUB231PD:
-            return e_vfmaddsub231pd;
-        case X86_INS_VFMADDSUB231PS:
-            return e_vfmaddsub231ps;
-        case X86_INS_VFMADDSUBPD:
-            return e_vfmaddsubpd;
-        case X86_INS_VFMADDSUBPS:
-            return e_vfmaddsubps;
-        case X86_INS_VFMSUB132PD:
-            return e_vfmsub132pd;
-        case X86_INS_VFMSUB132PS:
-            return e_vfmsub132ps;
-        case X86_INS_VFMSUB132SD:
-            return e_vfmsub132sd;
-        case X86_INS_VFMSUB132SS:
-            return e_vfmsub132ss;
-        case X86_INS_VFMSUB213PD:
-            return e_vfmsub213pd;
-        case X86_INS_VFMSUB213PS:
-            return e_vfmsub213ps;
-        case X86_INS_VFMSUB213SD:
-            return e_vfmsub213sd;
-        case X86_INS_VFMSUB213SS:
-            return e_vfmsub213ss;
-        case X86_INS_VFMSUB231PD:
-            return e_vfmsub231pd;
-        case X86_INS_VFMSUB231PS:
-            return e_vfmsub231ps;
-        case X86_INS_VFMSUB231SD:
-            return e_vfmsub231sd;
-        case X86_INS_VFMSUB231SS:
-            return e_vfmsub231ss;
-        case X86_INS_VFMSUBADD132PD:
-            return e_vfmsubadd132pd;
-        case X86_INS_VFMSUBADD132PS:
-            return e_vfmsubadd132ps;
-        case X86_INS_VFMSUBADD213PD:
-            return e_vfmsubadd213pd;
-        case X86_INS_VFMSUBADD213PS:
-            return e_vfmsubadd213ps;
-        case X86_INS_VFMSUBADD231PD:
-            return e_vfmsubadd231pd;
-        case X86_INS_VFMSUBADD231PS:
-            return e_vfmsubadd231ps;
-        case X86_INS_VFMSUBADDPD:
-            return e_vfmsubaddpd;
-        case X86_INS_VFMSUBADDPS:
-            return e_vfmsubaddps;
-        case X86_INS_VFMSUBPD:
-            return e_vfmsubpd;
-        case X86_INS_VFMSUBPS:
-            return e_vfmsubps;
-        case X86_INS_VFMSUBSD:
-            return e_vfmsubsd;
-        case X86_INS_VFMSUBSS:
-            return e_vfmsubss;
-        case X86_INS_VFNMADD132PD:
-            return e_vfnmadd132pd;
-        case X86_INS_VFNMADD132PS:
-            return e_vfnmadd132ps;
-        case X86_INS_VFNMADD132SD:
-            return e_vfnmadd132sd;
-        case X86_INS_VFNMADD132SS:
-            return e_vfnmadd132ss;
-        case X86_INS_VFNMADD213PD:
-            return e_vfnmadd213pd;
-        case X86_INS_VFNMADD213PS:
-            return e_vfnmadd213ps;
-        case X86_INS_VFNMADD213SD:
-            return e_vfnmadd213sd;
-        case X86_INS_VFNMADD213SS:
-            return e_vfnmadd213ss;
-        case X86_INS_VFNMADD231PD:
-            return e_vfnmadd231pd;
-        case X86_INS_VFNMADD231PS:
-            return e_vfnmadd231ps;
-        case X86_INS_VFNMADD231SD:
-            return e_vfnmadd231sd;
-        case X86_INS_VFNMADD231SS:
-            return e_vfnmadd231ss;
-        case X86_INS_VFNMADDPD:
-            return e_vfnmaddpd;
-        case X86_INS_VFNMADDPS:
-            return e_vfnmaddps;
-        case X86_INS_VFNMADDSD:
-            return e_vfnmaddsd;
-        case X86_INS_VFNMADDSS:
-            return e_vfnmaddss;
-        case X86_INS_VFNMSUB132PD:
-            return e_vfnmsub132pd;
-        case X86_INS_VFNMSUB132PS:
-            return e_vfnmsub132ps;
-        case X86_INS_VFNMSUB132SD:
-            return e_vfnmsub132sd;
-        case X86_INS_VFNMSUB132SS:
-            return e_vfnmsub132ss;
-        case X86_INS_VFNMSUB213PD:
-            return e_vfnmsub213pd;
-        case X86_INS_VFNMSUB213PS:
-            return e_vfnmsub213ps;
-        case X86_INS_VFNMSUB213SD:
-            return e_vfnmsub213sd;
-        case X86_INS_VFNMSUB213SS:
-            return e_vfnmsub213ss;
-        case X86_INS_VFNMSUB231PD:
-            return e_vfnmsub231pd;
-        case X86_INS_VFNMSUB231PS:
-            return e_vfnmsub231ps;
-        case X86_INS_VFNMSUB231SD:
-            return e_vfnmsub231sd;
-        case X86_INS_VFNMSUB231SS:
-            return e_vfnmsub231ss;
-        case X86_INS_VFNMSUBPD:
-            return e_vfnmsubpd;
-        case X86_INS_VFNMSUBPS:
-            return e_vfnmsubps;
-        case X86_INS_VFNMSUBSD:
-            return e_vfnmsubsd;
-        case X86_INS_VFNMSUBSS:
-            return e_vfnmsubss;
-        case X86_INS_VFPCLASSPD:
-            return e_vfpclasspd;
-        case X86_INS_VFPCLASSPS:
-            return e_vfpclassps;
-        case X86_INS_VFPCLASSSD:
-            return e_vfpclasssd;
-        case X86_INS_VFPCLASSSS:
-            return e_vfpclassss;
-        case X86_INS_VFRCZPD:
-            return e_vfrczpd;
-        case X86_INS_VFRCZPS:
-            return e_vfrczps;
-        case X86_INS_VFRCZSD:
-            return e_vfrczsd;
-        case X86_INS_VFRCZSS:
-            return e_vfrczss;
-        case X86_INS_VGATHERDPD:
-            return e_vgatherdpd;
-        case X86_INS_VGATHERDPS:
-            return e_vgatherdps;
-        case X86_INS_VGATHERPF0DPD:
-            return e_vgatherpf0dpd;
-        case X86_INS_VGATHERPF0DPS:
-            return e_vgatherpf0dps;
-        case X86_INS_VGATHERPF0QPD:
-            return e_vgatherpf0qpd;
-        case X86_INS_VGATHERPF0QPS:
-            return e_vgatherpf0qps;
-        case X86_INS_VGATHERPF1DPD:
-            return e_vgatherpf1dpd;
-        case X86_INS_VGATHERPF1DPS:
-            return e_vgatherpf1dps;
-        case X86_INS_VGATHERPF1QPD:
-            return e_vgatherpf1qpd;
-        case X86_INS_VGATHERPF1QPS:
-            return e_vgatherpf1qps;
-        case X86_INS_VGATHERQPD:
-            return e_vgatherqpd;
-        case X86_INS_VGATHERQPS:
-            return e_vgatherqps;
-        case X86_INS_VGETEXPPD:
-            return e_vgetexppd;
-        case X86_INS_VGETEXPPS:
-            return e_vgetexpps;
-        case X86_INS_VGETEXPSD:
-            return e_vgetexpsd;
-        case X86_INS_VGETEXPSS:
-            return e_vgetexpss;
-        case X86_INS_VGETMANTPD:
-            return e_vgetmantpd;
-        case X86_INS_VGETMANTPS:
-            return e_vgetmantps;
-        case X86_INS_VGETMANTSD:
-            return e_vgetmantsd;
-        case X86_INS_VGETMANTSS:
-            return e_vgetmantss;
-        case X86_INS_VGF2P8AFFINEINVQB:
-            return e_vgf2p8affineinvqb;
-        case X86_INS_VGF2P8AFFINEQB:
-            return e_vgf2p8affineqb;
-        case X86_INS_VGF2P8MULB:
-            return e_vgf2p8mulb;
-        case X86_INS_VHADDPD:
-            return e_vhaddpd;
-        case X86_INS_VHADDPS:
-            return e_vhaddps;
-        case X86_INS_VHSUBPD:
-            return e_vhsubpd;
-        case X86_INS_VHSUBPS:
-            return e_vhsubps;
-        case X86_INS_VINSERTF128:
-            return e_vinsertf128;
-        case X86_INS_VINSERTF32X4:
-            return e_vinsertf32x4;
-        case X86_INS_VINSERTF32X8:
-            return e_vinsertf32x8;
-        case X86_INS_VINSERTF64X2:
-            return e_vinsertf64x2;
-        case X86_INS_VINSERTF64X4:
-            return e_vinsertf64x4;
-        case X86_INS_VINSERTI128:
-            return e_vinserti128;
-        case X86_INS_VINSERTI32X4:
-            return e_vinserti32x4;
-        case X86_INS_VINSERTI32X8:
-            return e_vinserti32x8;
-        case X86_INS_VINSERTI64X2:
-            return e_vinserti64x2;
-        case X86_INS_VINSERTI64X4:
-            return e_vinserti64x4;
-        case X86_INS_VINSERTPS:
-            return e_vinsertps;
-        case X86_INS_VLDDQU:
-            return e_vlddqu;
-        case X86_INS_VLDMXCSR:
-            return e_vldmxcsr;
-        case X86_INS_VMASKMOVDQU:
-            return e_vmaskmovdqu;
-        case X86_INS_VMASKMOVPD:
-            return e_vmaskmovpd;
-        case X86_INS_VMASKMOVPS:
-            return e_vmaskmovps;
-        case X86_INS_VMAXPD:
-            return e_vmaxpd;
-        case X86_INS_VMAXPS:
-            return e_vmaxps;
-        case X86_INS_VMAXSD:
-            return e_vmaxsd;
-        case X86_INS_VMAXSS:
-            return e_vmaxss;
-        case X86_INS_VMCALL:
-            return e_vmcall;
-        case X86_INS_VMCLEAR:
-            return e_vmclear;
-        case X86_INS_VMFUNC:
-            return e_vmfunc;
-        case X86_INS_VMINPD:
-            return e_vminpd;
-        case X86_INS_VMINPS:
-            return e_vminps;
-        case X86_INS_VMINSD:
-            return e_vminsd;
-        case X86_INS_VMINSS:
-            return e_vminss;
-        case X86_INS_VMLAUNCH:
-            return e_vmlaunch;
-        case X86_INS_VMLOAD:
-            return e_vmload;
-        case X86_INS_VMMCALL:
-            return e_vmmcall;
-        case X86_INS_VMOVQ:
-            return e_vmovq;
-        case X86_INS_VMOVAPD:
-            return e_vmovapd;
-        case X86_INS_VMOVAPS:
-            return e_vmovaps;
-        case X86_INS_VMOVDDUP:
-            return e_vmovddup;
-        case X86_INS_VMOVD:
-            return e_vmovd;
-        case X86_INS_VMOVDQA32:
-            return e_vmovdqa32;
-        case X86_INS_VMOVDQA64:
-            return e_vmovdqa64;
-        case X86_INS_VMOVDQA:
-            return e_vmovdqa;
-        case X86_INS_VMOVDQU16:
-            return e_vmovdqu16;
-        case X86_INS_VMOVDQU32:
-            return e_vmovdqu32;
-        case X86_INS_VMOVDQU64:
-            return e_vmovdqu64;
-        case X86_INS_VMOVDQU8:
-            return e_vmovdqu8;
-        case X86_INS_VMOVDQU:
-            return e_vmovdqu;
-        case X86_INS_VMOVHLPS:
-            return e_vmovhlps;
-        case X86_INS_VMOVHPD:
-            return e_vmovhpd;
-        case X86_INS_VMOVHPS:
-            return e_vmovhps;
-        case X86_INS_VMOVLHPS:
-            return e_vmovlhps;
-        case X86_INS_VMOVLPD:
-            return e_vmovlpd;
-        case X86_INS_VMOVLPS:
-            return e_vmovlps;
-        case X86_INS_VMOVMSKPD:
-            return e_vmovmskpd;
-        case X86_INS_VMOVMSKPS:
-            return e_vmovmskps;
-        case X86_INS_VMOVNTDQA:
-            return e_vmovntdqa;
-        case X86_INS_VMOVNTDQ:
-            return e_vmovntdq;
-        case X86_INS_VMOVNTPD:
-            return e_vmovntpd;
-        case X86_INS_VMOVNTPS:
-            return e_vmovntps;
-        case X86_INS_VMOVSD:
-            return e_vmovsd;
-        case X86_INS_VMOVSHDUP:
-            return e_vmovshdup;
-        case X86_INS_VMOVSLDUP:
-            return e_vmovsldup;
-        case X86_INS_VMOVSS:
-            return e_vmovss;
-        case X86_INS_VMOVUPD:
-            return e_vmovupd;
-        case X86_INS_VMOVUPS:
-            return e_vmovups;
-        case X86_INS_VMPSADBW:
-            return e_vmpsadbw;
-        case X86_INS_VMPTRLD:
-            return e_vmptrld;
-        case X86_INS_VMPTRST:
-            return e_vmptrst;
-        case X86_INS_VMREAD:
-            return e_vmread;
-        case X86_INS_VMRESUME:
-            return e_vmresume;
-        case X86_INS_VMRUN:
-            return e_vmrun;
-        case X86_INS_VMSAVE:
-            return e_vmsave;
-        case X86_INS_VMULPD:
-            return e_vmulpd;
-        case X86_INS_VMULPS:
-            return e_vmulps;
-        case X86_INS_VMULSD:
-            return e_vmulsd;
-        case X86_INS_VMULSS:
-            return e_vmulss;
-        case X86_INS_VMWRITE:
-            return e_vmwrite;
-        case X86_INS_VMXOFF:
-            return e_vmxoff;
-        case X86_INS_VMXON:
-            return e_vmxon;
-        case X86_INS_VORPD:
-            return e_vorpd;
-        case X86_INS_VORPS:
-            return e_vorps;
-        case X86_INS_VP4DPWSSDS:
-            return e_vp4dpwssds;
-        case X86_INS_VP4DPWSSD:
-            return e_vp4dpwssd;
-        case X86_INS_VPABSB:
-            return e_vpabsb;
-        case X86_INS_VPABSD:
-            return e_vpabsd;
-        case X86_INS_VPABSQ:
-            return e_vpabsq;
-        case X86_INS_VPABSW:
-            return e_vpabsw;
-        case X86_INS_VPACKSSDW:
-            return e_vpackssdw;
-        case X86_INS_VPACKSSWB:
-            return e_vpacksswb;
-        case X86_INS_VPACKUSDW:
-            return e_vpackusdw;
-        case X86_INS_VPACKUSWB:
-            return e_vpackuswb;
-        case X86_INS_VPADDB:
-            return e_vpaddb;
-        case X86_INS_VPADDD:
-            return e_vpaddd;
-        case X86_INS_VPADDQ:
-            return e_vpaddq;
-        case X86_INS_VPADDSB:
-            return e_vpaddsb;
-        case X86_INS_VPADDSW:
-            return e_vpaddsw;
-        case X86_INS_VPADDUSB:
-            return e_vpaddusb;
-        case X86_INS_VPADDUSW:
-            return e_vpaddusw;
-        case X86_INS_VPADDW:
-            return e_vpaddw;
-        case X86_INS_VPALIGNR:
-            return e_vpalignr;
-        case X86_INS_VPANDD:
-            return e_vpandd;
-        case X86_INS_VPANDND:
-            return e_vpandnd;
-        case X86_INS_VPANDNQ:
-            return e_vpandnq;
-        case X86_INS_VPANDN:
-            return e_vpandn;
-        case X86_INS_VPANDQ:
-            return e_vpandq;
-        case X86_INS_VPAND:
-            return e_vpand;
-        case X86_INS_VPAVGB:
-            return e_vpavgb;
-        case X86_INS_VPAVGW:
-            return e_vpavgw;
-        case X86_INS_VPBLENDD:
-            return e_vpblendd;
-        case X86_INS_VPBLENDMB:
-            return e_vpblendmb;
-        case X86_INS_VPBLENDMD:
-            return e_vpblendmd;
-        case X86_INS_VPBLENDMQ:
-            return e_vpblendmq;
-        case X86_INS_VPBLENDMW:
-            return e_vpblendmw;
-        case X86_INS_VPBLENDVB:
-            return e_vpblendvb;
-        case X86_INS_VPBLENDW:
-            return e_vpblendw;
-        case X86_INS_VPBROADCASTB:
-            return e_vpbroadcastb;
-        case X86_INS_VPBROADCASTD:
-            return e_vpbroadcastd;
-        case X86_INS_VPBROADCASTMB2Q:
-            return e_vpbroadcastmb2q;
-        case X86_INS_VPBROADCASTMW2D:
-            return e_vpbroadcastmw2d;
-        case X86_INS_VPBROADCASTQ:
-            return e_vpbroadcastq;
-        case X86_INS_VPBROADCASTW:
-            return e_vpbroadcastw;
-        case X86_INS_VPCLMULQDQ:
-            return e_vpclmulqdq;
-        case X86_INS_VPCMOV:
-            return e_vpcmov;
-        case X86_INS_VPCMP:
-            return e_vpcmp;
-        case X86_INS_VPCMPB:
-            return e_vpcmpb;
-        case X86_INS_VPCMPD:
-            return e_vpcmpd;
-        case X86_INS_VPCMPEQB:
-            return e_vpcmpeqb;
-        case X86_INS_VPCMPEQD:
-            return e_vpcmpeqd;
-        case X86_INS_VPCMPEQQ:
-            return e_vpcmpeqq;
-        case X86_INS_VPCMPEQW:
-            return e_vpcmpeqw;
-        case X86_INS_VPCMPESTRI:
-            return e_vpcmpestri;
-        case X86_INS_VPCMPESTRM:
-            return e_vpcmpestrm;
-        case X86_INS_VPCMPGTB:
-            return e_vpcmpgtb;
-        case X86_INS_VPCMPGTD:
-            return e_vpcmpgtd;
-        case X86_INS_VPCMPGTQ:
-            return e_vpcmpgtq;
-        case X86_INS_VPCMPGTW:
-            return e_vpcmpgtw;
-        case X86_INS_VPCMPISTRI:
-            return e_vpcmpistri;
-        case X86_INS_VPCMPISTRM:
-            return e_vpcmpistrm;
-        case X86_INS_VPCMPQ:
-            return e_vpcmpq;
-        case X86_INS_VPCMPUB:
-            return e_vpcmpub;
-        case X86_INS_VPCMPUD:
-            return e_vpcmpud;
-        case X86_INS_VPCMPUQ:
-            return e_vpcmpuq;
-        case X86_INS_VPCMPUW:
-            return e_vpcmpuw;
-        case X86_INS_VPCMPW:
-            return e_vpcmpw;
-        case X86_INS_VPCOM:
-            return e_vpcom;
-        case X86_INS_VPCOMB:
-            return e_vpcomb;
-        case X86_INS_VPCOMD:
-            return e_vpcomd;
-        case X86_INS_VPCOMPRESSB:
-            return e_vpcompressb;
-        case X86_INS_VPCOMPRESSD:
-            return e_vpcompressd;
-        case X86_INS_VPCOMPRESSQ:
-            return e_vpcompressq;
-        case X86_INS_VPCOMPRESSW:
-            return e_vpcompressw;
-        case X86_INS_VPCOMQ:
-            return e_vpcomq;
-        case X86_INS_VPCOMUB:
-            return e_vpcomub;
-        case X86_INS_VPCOMUD:
-            return e_vpcomud;
-        case X86_INS_VPCOMUQ:
-            return e_vpcomuq;
-        case X86_INS_VPCOMUW:
-            return e_vpcomuw;
-        case X86_INS_VPCOMW:
-            return e_vpcomw;
-        case X86_INS_VPCONFLICTD:
-            return e_vpconflictd;
-        case X86_INS_VPCONFLICTQ:
-            return e_vpconflictq;
-        case X86_INS_VPDPBUSDS:
-            return e_vpdpbusds;
-        case X86_INS_VPDPBUSD:
-            return e_vpdpbusd;
-        case X86_INS_VPDPWSSDS:
-            return e_vpdpwssds;
-        case X86_INS_VPDPWSSD:
-            return e_vpdpwssd;
-        case X86_INS_VPERM2F128:
-            return e_vperm2f128;
-        case X86_INS_VPERM2I128:
-            return e_vperm2i128;
-        case X86_INS_VPERMB:
-            return e_vpermb;
-        case X86_INS_VPERMD:
-            return e_vpermd;
-        case X86_INS_VPERMI2B:
-            return e_vpermi2b;
-        case X86_INS_VPERMI2D:
-            return e_vpermi2d;
-        case X86_INS_VPERMI2PD:
-            return e_vpermi2pd;
-        case X86_INS_VPERMI2PS:
-            return e_vpermi2ps;
-        case X86_INS_VPERMI2Q:
-            return e_vpermi2q;
-        case X86_INS_VPERMI2W:
-            return e_vpermi2w;
-        case X86_INS_VPERMIL2PD:
-            return e_vpermil2pd;
-        case X86_INS_VPERMILPD:
-            return e_vpermilpd;
-        case X86_INS_VPERMIL2PS:
-            return e_vpermil2ps;
-        case X86_INS_VPERMILPS:
-            return e_vpermilps;
-        case X86_INS_VPERMPD:
-            return e_vpermpd;
-        case X86_INS_VPERMPS:
-            return e_vpermps;
-        case X86_INS_VPERMQ:
-            return e_vpermq;
-        case X86_INS_VPERMT2B:
-            return e_vpermt2b;
-        case X86_INS_VPERMT2D:
-            return e_vpermt2d;
-        case X86_INS_VPERMT2PD:
-            return e_vpermt2pd;
-        case X86_INS_VPERMT2PS:
-            return e_vpermt2ps;
-        case X86_INS_VPERMT2Q:
-            return e_vpermt2q;
-        case X86_INS_VPERMT2W:
-            return e_vpermt2w;
-        case X86_INS_VPERMW:
-            return e_vpermw;
-        case X86_INS_VPEXPANDB:
-            return e_vpexpandb;
-        case X86_INS_VPEXPANDD:
-            return e_vpexpandd;
-        case X86_INS_VPEXPANDQ:
-            return e_vpexpandq;
-        case X86_INS_VPEXPANDW:
-            return e_vpexpandw;
-        case X86_INS_VPEXTRB:
-            return e_vpextrb;
-        case X86_INS_VPEXTRD:
-            return e_vpextrd;
-        case X86_INS_VPEXTRQ:
-            return e_vpextrq;
-        case X86_INS_VPEXTRW:
-            return e_vpextrw;
-        case X86_INS_VPGATHERDD:
-            return e_vpgatherdd;
-        case X86_INS_VPGATHERDQ:
-            return e_vpgatherdq;
-        case X86_INS_VPGATHERQD:
-            return e_vpgatherqd;
-        case X86_INS_VPGATHERQQ:
-            return e_vpgatherqq;
-        case X86_INS_VPHADDBD:
-            return e_vphaddbd;
-        case X86_INS_VPHADDBQ:
-            return e_vphaddbq;
-        case X86_INS_VPHADDBW:
-            return e_vphaddbw;
-        case X86_INS_VPHADDDQ:
-            return e_vphadddq;
-        case X86_INS_VPHADDD:
-            return e_vphaddd;
-        case X86_INS_VPHADDSW:
-            return e_vphaddsw;
-        case X86_INS_VPHADDUBD:
-            return e_vphaddubd;
-        case X86_INS_VPHADDUBQ:
-            return e_vphaddubq;
-        case X86_INS_VPHADDUBW:
-            return e_vphaddubw;
-        case X86_INS_VPHADDUDQ:
-            return e_vphaddudq;
-        case X86_INS_VPHADDUWD:
-            return e_vphadduwd;
-        case X86_INS_VPHADDUWQ:
-            return e_vphadduwq;
-        case X86_INS_VPHADDWD:
-            return e_vphaddwd;
-        case X86_INS_VPHADDWQ:
-            return e_vphaddwq;
-        case X86_INS_VPHADDW:
-            return e_vphaddw;
-        case X86_INS_VPHMINPOSUW:
-            return e_vphminposuw;
-        case X86_INS_VPHSUBBW:
-            return e_vphsubbw;
-        case X86_INS_VPHSUBDQ:
-            return e_vphsubdq;
-        case X86_INS_VPHSUBD:
-            return e_vphsubd;
-        case X86_INS_VPHSUBSW:
-            return e_vphsubsw;
-        case X86_INS_VPHSUBWD:
-            return e_vphsubwd;
-        case X86_INS_VPHSUBW:
-            return e_vphsubw;
-        case X86_INS_VPINSRB:
-            return e_vpinsrb;
-        case X86_INS_VPINSRD:
-            return e_vpinsrd;
-        case X86_INS_VPINSRQ:
-            return e_vpinsrq;
-        case X86_INS_VPINSRW:
-            return e_vpinsrw;
-        case X86_INS_VPLZCNTD:
-            return e_vplzcntd;
-        case X86_INS_VPLZCNTQ:
-            return e_vplzcntq;
-        case X86_INS_VPMACSDD:
-            return e_vpmacsdd;
-        case X86_INS_VPMACSDQH:
-            return e_vpmacsdqh;
-        case X86_INS_VPMACSDQL:
-            return e_vpmacsdql;
-        case X86_INS_VPMACSSDD:
-            return e_vpmacssdd;
-        case X86_INS_VPMACSSDQH:
-            return e_vpmacssdqh;
-        case X86_INS_VPMACSSDQL:
-            return e_vpmacssdql;
-        case X86_INS_VPMACSSWD:
-            return e_vpmacsswd;
-        case X86_INS_VPMACSSWW:
-            return e_vpmacssww;
-        case X86_INS_VPMACSWD:
-            return e_vpmacswd;
-        case X86_INS_VPMACSWW:
-            return e_vpmacsww;
-        case X86_INS_VPMADCSSWD:
-            return e_vpmadcsswd;
-        case X86_INS_VPMADCSWD:
-            return e_vpmadcswd;
-        case X86_INS_VPMADD52HUQ:
-            return e_vpmadd52huq;
-        case X86_INS_VPMADD52LUQ:
-            return e_vpmadd52luq;
-        case X86_INS_VPMADDUBSW:
-            return e_vpmaddubsw;
-        case X86_INS_VPMADDWD:
-            return e_vpmaddwd;
-        case X86_INS_VPMASKMOVD:
-            return e_vpmaskmovd;
-        case X86_INS_VPMASKMOVQ:
-            return e_vpmaskmovq;
-        case X86_INS_VPMAXSB:
-            return e_vpmaxsb;
-        case X86_INS_VPMAXSD:
-            return e_vpmaxsd;
-        case X86_INS_VPMAXSQ:
-            return e_vpmaxsq;
-        case X86_INS_VPMAXSW:
-            return e_vpmaxsw;
-        case X86_INS_VPMAXUB:
-            return e_vpmaxub;
-        case X86_INS_VPMAXUD:
-            return e_vpmaxud;
-        case X86_INS_VPMAXUQ:
-            return e_vpmaxuq;
-        case X86_INS_VPMAXUW:
-            return e_vpmaxuw;
-        case X86_INS_VPMINSB:
-            return e_vpminsb;
-        case X86_INS_VPMINSD:
-            return e_vpminsd;
-        case X86_INS_VPMINSQ:
-            return e_vpminsq;
-        case X86_INS_VPMINSW:
-            return e_vpminsw;
-        case X86_INS_VPMINUB:
-            return e_vpminub;
-        case X86_INS_VPMINUD:
-            return e_vpminud;
-        case X86_INS_VPMINUQ:
-            return e_vpminuq;
-        case X86_INS_VPMINUW:
-            return e_vpminuw;
-        case X86_INS_VPMOVB2M:
-            return e_vpmovb2m;
-        case X86_INS_VPMOVD2M:
-            return e_vpmovd2m;
-        case X86_INS_VPMOVDB:
-            return e_vpmovdb;
-        case X86_INS_VPMOVDW:
-            return e_vpmovdw;
-        case X86_INS_VPMOVM2B:
-            return e_vpmovm2b;
-        case X86_INS_VPMOVM2D:
-            return e_vpmovm2d;
-        case X86_INS_VPMOVM2Q:
-            return e_vpmovm2q;
-        case X86_INS_VPMOVM2W:
-            return e_vpmovm2w;
-        case X86_INS_VPMOVMSKB:
-            return e_vpmovmskb;
-        case X86_INS_VPMOVQ2M:
-            return e_vpmovq2m;
-        case X86_INS_VPMOVQB:
-            return e_vpmovqb;
-        case X86_INS_VPMOVQD:
-            return e_vpmovqd;
-        case X86_INS_VPMOVQW:
-            return e_vpmovqw;
-        case X86_INS_VPMOVSDB:
-            return e_vpmovsdb;
-        case X86_INS_VPMOVSDW:
-            return e_vpmovsdw;
-        case X86_INS_VPMOVSQB:
-            return e_vpmovsqb;
-        case X86_INS_VPMOVSQD:
-            return e_vpmovsqd;
-        case X86_INS_VPMOVSQW:
-            return e_vpmovsqw;
-        case X86_INS_VPMOVSWB:
-            return e_vpmovswb;
-        case X86_INS_VPMOVSXBD:
-            return e_vpmovsxbd;
-        case X86_INS_VPMOVSXBQ:
-            return e_vpmovsxbq;
-        case X86_INS_VPMOVSXBW:
-            return e_vpmovsxbw;
-        case X86_INS_VPMOVSXDQ:
-            return e_vpmovsxdq;
-        case X86_INS_VPMOVSXWD:
-            return e_vpmovsxwd;
-        case X86_INS_VPMOVSXWQ:
-            return e_vpmovsxwq;
-        case X86_INS_VPMOVUSDB:
-            return e_vpmovusdb;
-        case X86_INS_VPMOVUSDW:
-            return e_vpmovusdw;
-        case X86_INS_VPMOVUSQB:
-            return e_vpmovusqb;
-        case X86_INS_VPMOVUSQD:
-            return e_vpmovusqd;
-        case X86_INS_VPMOVUSQW:
-            return e_vpmovusqw;
-        case X86_INS_VPMOVUSWB:
-            return e_vpmovuswb;
-        case X86_INS_VPMOVW2M:
-            return e_vpmovw2m;
-        case X86_INS_VPMOVWB:
-            return e_vpmovwb;
-        case X86_INS_VPMOVZXBD:
-            return e_vpmovzxbd;
-        case X86_INS_VPMOVZXBQ:
-            return e_vpmovzxbq;
-        case X86_INS_VPMOVZXBW:
-            return e_vpmovzxbw;
-        case X86_INS_VPMOVZXDQ:
-            return e_vpmovzxdq;
-        case X86_INS_VPMOVZXWD:
-            return e_vpmovzxwd;
-        case X86_INS_VPMOVZXWQ:
-            return e_vpmovzxwq;
-        case X86_INS_VPMULDQ:
-            return e_vpmuldq;
-        case X86_INS_VPMULHRSW:
-            return e_vpmulhrsw;
-        case X86_INS_VPMULHUW:
-            return e_vpmulhuw;
-        case X86_INS_VPMULHW:
-            return e_vpmulhw;
-        case X86_INS_VPMULLD:
-            return e_vpmulld;
-        case X86_INS_VPMULLQ:
-            return e_vpmullq;
-        case X86_INS_VPMULLW:
-            return e_vpmullw;
-        case X86_INS_VPMULTISHIFTQB:
-            return e_vpmultishiftqb;
-        case X86_INS_VPMULUDQ:
-            return e_vpmuludq;
-        case X86_INS_VPOPCNTB:
-            return e_vpopcntb;
-        case X86_INS_VPOPCNTD:
-            return e_vpopcntd;
-        case X86_INS_VPOPCNTQ:
-            return e_vpopcntq;
-        case X86_INS_VPOPCNTW:
-            return e_vpopcntw;
-        case X86_INS_VPORD:
-            return e_vpord;
-        case X86_INS_VPORQ:
-            return e_vporq;
-        case X86_INS_VPOR:
-            return e_vpor;
-        case X86_INS_VPPERM:
-            return e_vpperm;
-        case X86_INS_VPROLD:
-            return e_vprold;
-        case X86_INS_VPROLQ:
-            return e_vprolq;
-        case X86_INS_VPROLVD:
-            return e_vprolvd;
-        case X86_INS_VPROLVQ:
-            return e_vprolvq;
-        case X86_INS_VPRORD:
-            return e_vprord;
-        case X86_INS_VPRORQ:
-            return e_vprorq;
-        case X86_INS_VPRORVD:
-            return e_vprorvd;
-        case X86_INS_VPRORVQ:
-            return e_vprorvq;
-        case X86_INS_VPROTB:
-            return e_vprotb;
-        case X86_INS_VPROTD:
-            return e_vprotd;
-        case X86_INS_VPROTQ:
-            return e_vprotq;
-        case X86_INS_VPROTW:
-            return e_vprotw;
-        case X86_INS_VPSADBW:
-            return e_vpsadbw;
-        case X86_INS_VPSCATTERDD:
-            return e_vpscatterdd;
-        case X86_INS_VPSCATTERDQ:
-            return e_vpscatterdq;
-        case X86_INS_VPSCATTERQD:
-            return e_vpscatterqd;
-        case X86_INS_VPSCATTERQQ:
-            return e_vpscatterqq;
-        case X86_INS_VPSHAB:
-            return e_vpshab;
-        case X86_INS_VPSHAD:
-            return e_vpshad;
-        case X86_INS_VPSHAQ:
-            return e_vpshaq;
-        case X86_INS_VPSHAW:
-            return e_vpshaw;
-        case X86_INS_VPSHLB:
-            return e_vpshlb;
-        case X86_INS_VPSHLDD:
-            return e_vpshldd;
-        case X86_INS_VPSHLDQ:
-            return e_vpshldq;
-        case X86_INS_VPSHLDVD:
-            return e_vpshldvd;
-        case X86_INS_VPSHLDVQ:
-            return e_vpshldvq;
-        case X86_INS_VPSHLDVW:
-            return e_vpshldvw;
-        case X86_INS_VPSHLDW:
-            return e_vpshldw;
-        case X86_INS_VPSHLD:
-            return e_vpshld;
-        case X86_INS_VPSHLQ:
-            return e_vpshlq;
-        case X86_INS_VPSHLW:
-            return e_vpshlw;
-        case X86_INS_VPSHRDD:
-            return e_vpshrdd;
-        case X86_INS_VPSHRDQ:
-            return e_vpshrdq;
-        case X86_INS_VPSHRDVD:
-            return e_vpshrdvd;
-        case X86_INS_VPSHRDVQ:
-            return e_vpshrdvq;
-        case X86_INS_VPSHRDVW:
-            return e_vpshrdvw;
-        case X86_INS_VPSHRDW:
-            return e_vpshrdw;
-        case X86_INS_VPSHUFBITQMB:
-            return e_vpshufbitqmb;
-        case X86_INS_VPSHUFB:
-            return e_vpshufb;
-        case X86_INS_VPSHUFD:
-            return e_vpshufd;
-        case X86_INS_VPSHUFHW:
-            return e_vpshufhw;
-        case X86_INS_VPSHUFLW:
-            return e_vpshuflw;
-        case X86_INS_VPSIGNB:
-            return e_vpsignb;
-        case X86_INS_VPSIGND:
-            return e_vpsignd;
-        case X86_INS_VPSIGNW:
-            return e_vpsignw;
-        case X86_INS_VPSLLDQ:
-            return e_vpslldq;
-        case X86_INS_VPSLLD:
-            return e_vpslld;
-        case X86_INS_VPSLLQ:
-            return e_vpsllq;
-        case X86_INS_VPSLLVD:
-            return e_vpsllvd;
-        case X86_INS_VPSLLVQ:
-            return e_vpsllvq;
-        case X86_INS_VPSLLVW:
-            return e_vpsllvw;
-        case X86_INS_VPSLLW:
-            return e_vpsllw;
-        case X86_INS_VPSRAD:
-            return e_vpsrad;
-        case X86_INS_VPSRAQ:
-            return e_vpsraq;
-        case X86_INS_VPSRAVD:
-            return e_vpsravd;
-        case X86_INS_VPSRAVQ:
-            return e_vpsravq;
-        case X86_INS_VPSRAVW:
-            return e_vpsravw;
-        case X86_INS_VPSRAW:
-            return e_vpsraw;
-        case X86_INS_VPSRLDQ:
-            return e_vpsrldq;
-        case X86_INS_VPSRLD:
-            return e_vpsrld;
-        case X86_INS_VPSRLQ:
-            return e_vpsrlq;
-        case X86_INS_VPSRLVD:
-            return e_vpsrlvd;
-        case X86_INS_VPSRLVQ:
-            return e_vpsrlvq;
-        case X86_INS_VPSRLVW:
-            return e_vpsrlvw;
-        case X86_INS_VPSRLW:
-            return e_vpsrlw;
-        case X86_INS_VPSUBB:
-            return e_vpsubb;
-        case X86_INS_VPSUBD:
-            return e_vpsubd;
-        case X86_INS_VPSUBQ:
-            return e_vpsubq;
-        case X86_INS_VPSUBSB:
-            return e_vpsubsb;
-        case X86_INS_VPSUBSW:
-            return e_vpsubsw;
-        case X86_INS_VPSUBUSB:
-            return e_vpsubusb;
-        case X86_INS_VPSUBUSW:
-            return e_vpsubusw;
-        case X86_INS_VPSUBW:
-            return e_vpsubw;
-        case X86_INS_VPTERNLOGD:
-            return e_vpternlogd;
-        case X86_INS_VPTERNLOGQ:
-            return e_vpternlogq;
-        case X86_INS_VPTESTMB:
-            return e_vptestmb;
-        case X86_INS_VPTESTMD:
-            return e_vptestmd;
-        case X86_INS_VPTESTMQ:
-            return e_vptestmq;
-        case X86_INS_VPTESTMW:
-            return e_vptestmw;
-        case X86_INS_VPTESTNMB:
-            return e_vptestnmb;
-        case X86_INS_VPTESTNMD:
-            return e_vptestnmd;
-        case X86_INS_VPTESTNMQ:
-            return e_vptestnmq;
-        case X86_INS_VPTESTNMW:
-            return e_vptestnmw;
-        case X86_INS_VPTEST:
-            return e_vptest;
-        case X86_INS_VPUNPCKHBW:
-            return e_vpunpckhbw;
-        case X86_INS_VPUNPCKHDQ:
-            return e_vpunpckhdq;
-        case X86_INS_VPUNPCKHQDQ:
-            return e_vpunpckhqdq;
-        case X86_INS_VPUNPCKHWD:
-            return e_vpunpckhwd;
-        case X86_INS_VPUNPCKLBW:
-            return e_vpunpcklbw;
-        case X86_INS_VPUNPCKLDQ:
-            return e_vpunpckldq;
-        case X86_INS_VPUNPCKLQDQ:
-            return e_vpunpcklqdq;
-        case X86_INS_VPUNPCKLWD:
-            return e_vpunpcklwd;
-        case X86_INS_VPXORD:
-            return e_vpxord;
-        case X86_INS_VPXORQ:
-            return e_vpxorq;
-        case X86_INS_VPXOR:
-            return e_vpxor;
-        case X86_INS_VRANGEPD:
-            return e_vrangepd;
-        case X86_INS_VRANGEPS:
-            return e_vrangeps;
-        case X86_INS_VRANGESD:
-            return e_vrangesd;
-        case X86_INS_VRANGESS:
-            return e_vrangess;
-        case X86_INS_VRCP14PD:
-            return e_vrcp14pd;
-        case X86_INS_VRCP14PS:
-            return e_vrcp14ps;
-        case X86_INS_VRCP14SD:
-            return e_vrcp14sd;
-        case X86_INS_VRCP14SS:
-            return e_vrcp14ss;
-        case X86_INS_VRCP28PD:
-            return e_vrcp28pd;
-        case X86_INS_VRCP28PS:
-            return e_vrcp28ps;
-        case X86_INS_VRCP28SD:
-            return e_vrcp28sd;
-        case X86_INS_VRCP28SS:
-            return e_vrcp28ss;
-        case X86_INS_VRCPPS:
-            return e_vrcpps;
-        case X86_INS_VRCPSS:
-            return e_vrcpss;
-        case X86_INS_VREDUCEPD:
-            return e_vreducepd;
-        case X86_INS_VREDUCEPS:
-            return e_vreduceps;
-        case X86_INS_VREDUCESD:
-            return e_vreducesd;
-        case X86_INS_VREDUCESS:
-            return e_vreducess;
-        case X86_INS_VRNDSCALEPD:
-            return e_vrndscalepd;
-        case X86_INS_VRNDSCALEPS:
-            return e_vrndscaleps;
-        case X86_INS_VRNDSCALESD:
-            return e_vrndscalesd;
-        case X86_INS_VRNDSCALESS:
-            return e_vrndscaless;
-        case X86_INS_VROUNDPD:
-            return e_vroundpd;
-        case X86_INS_VROUNDPS:
-            return e_vroundps;
-        case X86_INS_VROUNDSD:
-            return e_vroundsd;
-        case X86_INS_VROUNDSS:
-            return e_vroundss;
-        case X86_INS_VRSQRT14PD:
-            return e_vrsqrt14pd;
-        case X86_INS_VRSQRT14PS:
-            return e_vrsqrt14ps;
-        case X86_INS_VRSQRT14SD:
-            return e_vrsqrt14sd;
-        case X86_INS_VRSQRT14SS:
-            return e_vrsqrt14ss;
-        case X86_INS_VRSQRT28PD:
-            return e_vrsqrt28pd;
-        case X86_INS_VRSQRT28PS:
-            return e_vrsqrt28ps;
-        case X86_INS_VRSQRT28SD:
-            return e_vrsqrt28sd;
-        case X86_INS_VRSQRT28SS:
-            return e_vrsqrt28ss;
-        case X86_INS_VRSQRTPS:
-            return e_vrsqrtps;
-        case X86_INS_VRSQRTSS:
-            return e_vrsqrtss;
-        case X86_INS_VSCALEFPD:
-            return e_vscalefpd;
-        case X86_INS_VSCALEFPS:
-            return e_vscalefps;
-        case X86_INS_VSCALEFSD:
-            return e_vscalefsd;
-        case X86_INS_VSCALEFSS:
-            return e_vscalefss;
-        case X86_INS_VSCATTERDPD:
-            return e_vscatterdpd;
-        case X86_INS_VSCATTERDPS:
-            return e_vscatterdps;
-        case X86_INS_VSCATTERPF0DPD:
-            return e_vscatterpf0dpd;
-        case X86_INS_VSCATTERPF0DPS:
-            return e_vscatterpf0dps;
-        case X86_INS_VSCATTERPF0QPD:
-            return e_vscatterpf0qpd;
-        case X86_INS_VSCATTERPF0QPS:
-            return e_vscatterpf0qps;
-        case X86_INS_VSCATTERPF1DPD:
-            return e_vscatterpf1dpd;
-        case X86_INS_VSCATTERPF1DPS:
-            return e_vscatterpf1dps;
-        case X86_INS_VSCATTERPF1QPD:
-            return e_vscatterpf1qpd;
-        case X86_INS_VSCATTERPF1QPS:
-            return e_vscatterpf1qps;
-        case X86_INS_VSCATTERQPD:
-            return e_vscatterqpd;
-        case X86_INS_VSCATTERQPS:
-            return e_vscatterqps;
-        case X86_INS_VSHUFF32X4:
-            return e_vshuff32x4;
-        case X86_INS_VSHUFF64X2:
-            return e_vshuff64x2;
-        case X86_INS_VSHUFI32X4:
-            return e_vshufi32x4;
-        case X86_INS_VSHUFI64X2:
-            return e_vshufi64x2;
-        case X86_INS_VSHUFPD:
-            return e_vshufpd;
-        case X86_INS_VSHUFPS:
-            return e_vshufps;
-        case X86_INS_VSQRTPD:
-            return e_vsqrtpd;
-        case X86_INS_VSQRTPS:
-            return e_vsqrtps;
-        case X86_INS_VSQRTSD:
-            return e_vsqrtsd;
-        case X86_INS_VSQRTSS:
-            return e_vsqrtss;
-        case X86_INS_VSTMXCSR:
-            return e_vstmxcsr;
-        case X86_INS_VSUBPD:
-            return e_vsubpd;
-        case X86_INS_VSUBPS:
-            return e_vsubps;
-        case X86_INS_VSUBSD:
-            return e_vsubsd;
-        case X86_INS_VSUBSS:
-            return e_vsubss;
-        case X86_INS_VTESTPD:
-            return e_vtestpd;
-        case X86_INS_VTESTPS:
-            return e_vtestps;
-        case X86_INS_VUCOMISD:
-            return e_vucomisd;
-        case X86_INS_VUCOMISS:
-            return e_vucomiss;
-        case X86_INS_VUNPCKHPD:
-            return e_vunpckhpd;
-        case X86_INS_VUNPCKHPS:
-            return e_vunpckhps;
-        case X86_INS_VUNPCKLPD:
-            return e_vunpcklpd;
-        case X86_INS_VUNPCKLPS:
-            return e_vunpcklps;
-        case X86_INS_VXORPD:
-            return e_vxorpd;
-        case X86_INS_VXORPS:
-            return e_vxorps;
-        case X86_INS_VZEROALL:
-            return e_vzeroall;
-        case X86_INS_VZEROUPPER:
-            return e_vzeroupper;
-        case X86_INS_WAIT:
-            return e_wait;
-        case X86_INS_WBINVD:
-            return e_wbinvd;
-        case X86_INS_WBNOINVD:
-            return e_wbnoinvd;
-        case X86_INS_WRFSBASE:
-            return e_wrfsbase;
-        case X86_INS_WRGSBASE:
-            return e_wrgsbase;
-        case X86_INS_WRMSR:
-            return e_wrmsr;
-        case X86_INS_WRPKRU:
-            return e_wrpkru;
-        case X86_INS_WRSSD:
-            return e_wrssd;
-        case X86_INS_WRSSQ:
-            return e_wrssq;
-        case X86_INS_WRUSSD:
-            return e_wrussd;
-        case X86_INS_WRUSSQ:
-            return e_wrussq;
-        case X86_INS_XABORT:
-            return e_xabort;
-        case X86_INS_XACQUIRE:
-            return e_xacquire;
-        case X86_INS_XADD:
-            return e_xadd;
-        case X86_INS_XBEGIN:
-            return e_xbegin;
-        case X86_INS_XCHG:
-            return e_xchg;
-        case X86_INS_FXCH:
-            return e_fxch;
-        case X86_INS_XCRYPTCBC:
-            return e_xcryptcbc;
-        case X86_INS_XCRYPTCFB:
-            return e_xcryptcfb;
-        case X86_INS_XCRYPTCTR:
-            return e_xcryptctr;
-        case X86_INS_XCRYPTECB:
-            return e_xcryptecb;
-        case X86_INS_XCRYPTOFB:
-            return e_xcryptofb;
-        case X86_INS_XEND:
-            return e_xend;
-        case X86_INS_XGETBV:
-            return e_xgetbv;
-        case X86_INS_XLATB:
-            return e_xlatb;
-        case X86_INS_XOR:
-            return e_xor;
-        case X86_INS_XORPD:
-            return e_xorpd;
-        case X86_INS_XORPS:
-            return e_xorps;
-        case X86_INS_XRELEASE:
-            return e_xrelease;
-        case X86_INS_XRSTOR:
-            return e_xrstor;
-        case X86_INS_XRSTOR64:
-            return e_xrstor64;
-        case X86_INS_XRSTORS:
-            return e_xrstors;
-        case X86_INS_XRSTORS64:
-            return e_xrstors64;
-        case X86_INS_XSAVE:
-            return e_xsave;
-        case X86_INS_XSAVE64:
-            return e_xsave64;
-        case X86_INS_XSAVEC:
-            return e_xsavec;
-        case X86_INS_XSAVEC64:
-            return e_xsavec64;
-        case X86_INS_XSAVEOPT:
-            return e_xsaveopt;
-        case X86_INS_XSAVEOPT64:
-            return e_xsaveopt64;
-        case X86_INS_XSAVES:
-            return e_xsaves;
-        case X86_INS_XSAVES64:
-            return e_xsaves64;
-        case X86_INS_XSETBV:
-            return e_xsetbv;
-        case X86_INS_XSHA1:
-            return e_xsha1;
-        case X86_INS_XSHA256:
-            return e_xsha256;
-        case X86_INS_XSTORE:
-            return e_xstore;
-        case X86_INS_XTEST:
-            return e_xtest;
-        default:
-            return e_No_Entry;
-    }
-
-}
-
-MachRegister InstructionDecoder_Capstone::registerTranslation_x86(x86_reg cap_reg) {
-    switch (cap_reg) {
-        case X86_REG_AH:
-            return x86_64::ah;
-        case X86_REG_AL:
-            return x86_64::al;
-        case X86_REG_AX:
-            return x86_64::ax;
-        case X86_REG_BH:
-            return x86_64::bh;
-        case X86_REG_BL:
-            return x86_64::bl;
-        case X86_REG_BP:
-            return x86_64::bp;
-        case X86_REG_BPL:
-            return x86_64::bpl;
-        case X86_REG_BX:
-            return x86_64::bx;
-        case X86_REG_CH:
-            return x86_64::ch;
-        case X86_REG_CL:
-            return x86_64::cl;
-        case X86_REG_CS:
-            return x86_64::cs;
-        case X86_REG_CX:
-            return x86_64::cx;
-        case X86_REG_DH:
-            return x86_64::dh;
-        case X86_REG_DI:
-            return x86_64::di;
-        case X86_REG_DIL:
-            return x86_64::dil;
-        case X86_REG_DL:
-            return x86_64::dl;
-        case X86_REG_DS:
-            return x86_64::ds;
-        case X86_REG_DX:
-            return x86_64::dx;
-        case X86_REG_EAX:
-            return x86_64::eax;
-        case X86_REG_EBP:
-            return x86_64::ebp;
-        case X86_REG_EBX:
-            return x86_64::ebx;
-        case X86_REG_ECX:
-            return x86_64::ecx;
-        case X86_REG_EDI:
-            return x86_64::edi;
-        case X86_REG_EDX:
-            return x86_64::edx;
-        case X86_REG_EFLAGS:
-            return x86_64::flags;
-        case X86_REG_EIP:
-            return x86_64::eip;
-//        case X86_REG_EIZ:
-//            return x86_64::eiz;
-        case X86_REG_ES:
-            return x86_64::es;
-        case X86_REG_ESI:
-            return x86_64::esi;
-        case X86_REG_ESP:
-            return x86_64::esp;
-//        case X86_REG_FPSW:
-//            return x86_64::fpsw;
-        case X86_REG_FS:
-            return x86_64::fs;
-        case X86_REG_GS:
-            return x86_64::gs;
-        case X86_REG_IP:
-            return x86_64::rip;
-        case X86_REG_RAX:
-            return x86_64::rax;
-        case X86_REG_RBP:
-            return x86_64::rbp;
-        case X86_REG_RBX:
-            return x86_64::rbx;
-        case X86_REG_RCX:
-            return x86_64::rcx;
-        case X86_REG_RDI:
-            return x86_64::rdi;
-        case X86_REG_RDX:
-            return x86_64::rdx;
-        case X86_REG_RIP:
-            return x86_64::rip;
-//        case X86_REG_RIZ:
-//            return x86_64::riz;
-        case X86_REG_RSI:
-            return x86_64::rsi;
-        case X86_REG_RSP:
-            return x86_64::rsp;
-        case X86_REG_SI:
-            return x86_64::si;
-        case X86_REG_SIL:
-            return x86_64::sil;
-        case X86_REG_SP:
-            return x86_64::sp;
-        case X86_REG_SPL:
-            return x86_64::spl;
-        case X86_REG_SS:
-            return x86_64::ss;
-        case X86_REG_CR0:
-            return x86_64::cr0;
-        case X86_REG_CR1:
-            return x86_64::cr1;
-        case X86_REG_CR2:
-            return x86_64::cr2;
-        case X86_REG_CR3:
-            return x86_64::cr3;
-        case X86_REG_CR4:
-            return x86_64::cr4;
-        case X86_REG_CR5:
-            return x86_64::cr5;
-        case X86_REG_CR6:
-            return x86_64::cr6;
-        case X86_REG_CR7:
-            return x86_64::cr7;
-      /*
-        case X86_REG_CR8:
-            return x86_64::cr8;
-        case X86_REG_CR9:
-            return x86_64::cr9;
-        case X86_REG_CR10:
-            return x86_64::cr10;
-        case X86_REG_CR11:
-            return x86_64::cr11;
-        case X86_REG_CR12:
-            return x86_64::cr12;
-        case X86_REG_CR13:
-            return x86_64::cr13;
-        case X86_REG_CR14:
-            return x86_64::cr14;
-        case X86_REG_CR15:
-            return x86_64::cr15;
-       */
-        case X86_REG_DR0:
-            return x86_64::dr0;
-        case X86_REG_DR1:
-            return x86_64::dr1;
-        case X86_REG_DR2:
-            return x86_64::dr2;
-        case X86_REG_DR3:
-            return x86_64::dr3;
-        case X86_REG_DR4:
-            return x86_64::dr4;
-        case X86_REG_DR5:
-            return x86_64::dr5;
-        case X86_REG_DR6:
-            return x86_64::dr6;
-        case X86_REG_DR7:
-            return x86_64::dr7;
-       /*
-        case X86_REG_DR8:
-            return x86_64::dr8;
-        case X86_REG_DR9:
-            return x86_64::dr9;
-        case X86_REG_DR10:
-            return x86_64::dr10;
-        case X86_REG_DR11:
-            return x86_64::dr11;
-        case X86_REG_DR12:
-            return x86_64::dr12;
-        case X86_REG_DR13:
-            return x86_64::dr13;
-        case X86_REG_DR14:
-            return x86_64::dr14;
-        case X86_REG_DR15:
-            return x86_64::dr15;
-        case X86_REG_FP0:
-            return x86_64::fp0;
-        case X86_REG_FP1:
-            return x86_64::fp1;
-        case X86_REG_FP2:
-            return x86_64::fp2;
-        case X86_REG_FP3:
-            return x86_64::fp3;
-        case X86_REG_FP4:
-            return x86_64::fp4;
-        case X86_REG_FP5:
-            return x86_64::fp5;
-        case X86_REG_FP6:
-            return x86_64::fp6;
-        case X86_REG_FP7:
-            return x86_64::fp7;
-       */
-        case X86_REG_K0:
-            return x86_64::k0;
-        case X86_REG_K1:
-            return x86_64::k1;
-        case X86_REG_K2:
-            return x86_64::k2;
-        case X86_REG_K3:
-            return x86_64::k3;
-        case X86_REG_K4:
-            return x86_64::k4;
-        case X86_REG_K5:
-            return x86_64::k5;
-        case X86_REG_K6:
-            return x86_64::k6;
-        case X86_REG_K7:
-            return x86_64::k7;
-        case X86_REG_MM0:
-            return x86_64::mm0;
-        case X86_REG_MM1:
-            return x86_64::mm1;
-        case X86_REG_MM2:
-            return x86_64::mm2;
-        case X86_REG_MM3:
-            return x86_64::mm3;
-        case X86_REG_MM4:
-            return x86_64::mm4;
-        case X86_REG_MM5:
-            return x86_64::mm5;
-        case X86_REG_MM6:
-            return x86_64::mm6;
-        case X86_REG_MM7:
-            return x86_64::mm7;
-        case X86_REG_R8:
-            return x86_64::r8;
-        case X86_REG_R9:
-            return x86_64::r9;
-        case X86_REG_R10:
-            return x86_64::r10;
-        case X86_REG_R11:
-            return x86_64::r11;
-        case X86_REG_R12:
-            return x86_64::r12;
-        case X86_REG_R13:
-            return x86_64::r13;
-        case X86_REG_R14:
-            return x86_64::r14;
-        case X86_REG_R15:
-            return x86_64::r15;
-        case X86_REG_ST0:
-            return x86_64::st0;
-        case X86_REG_ST1:
-            return x86_64::st1;
-        case X86_REG_ST2:
-            return x86_64::st2;
-        case X86_REG_ST3:
-            return x86_64::st3;
-        case X86_REG_ST4:
-            return x86_64::st4;
-        case X86_REG_ST5:
-            return x86_64::st5;
-        case X86_REG_ST6:
-            return x86_64::st6;
-        case X86_REG_ST7:
-            return x86_64::st7;
-        case X86_REG_XMM0:
-            return x86_64::xmm0;
-        case X86_REG_XMM1:
-            return x86_64::xmm1;
-        case X86_REG_XMM2:
-            return x86_64::xmm2;
-        case X86_REG_XMM3:
-            return x86_64::xmm3;
-        case X86_REG_XMM4:
-            return x86_64::xmm4;
-        case X86_REG_XMM5:
-            return x86_64::xmm5;
-        case X86_REG_XMM6:
-            return x86_64::xmm6;
-        case X86_REG_XMM7:
-            return x86_64::xmm7;
-        case X86_REG_XMM8:
-            return x86_64::xmm8;
-        case X86_REG_XMM9:
-            return x86_64::xmm9;
-        case X86_REG_XMM10:
-            return x86_64::xmm10;
-        case X86_REG_XMM11:
-            return x86_64::xmm11;
-        case X86_REG_XMM12:
-            return x86_64::xmm12;
-        case X86_REG_XMM13:
-            return x86_64::xmm13;
-        case X86_REG_XMM14:
-            return x86_64::xmm14;
-        case X86_REG_XMM15:
-            return x86_64::xmm15;
-        case X86_REG_XMM16:
-            return x86_64::xmm16;
-        case X86_REG_XMM17:
-            return x86_64::xmm17;
-        case X86_REG_XMM18:
-            return x86_64::xmm18;
-        case X86_REG_XMM19:
-            return x86_64::xmm19;
-        case X86_REG_XMM20:
-            return x86_64::xmm20;
-        case X86_REG_XMM21:
-            return x86_64::xmm21;
-        case X86_REG_XMM22:
-            return x86_64::xmm22;
-        case X86_REG_XMM23:
-            return x86_64::xmm23;
-        case X86_REG_XMM24:
-            return x86_64::xmm24;
-        case X86_REG_XMM25:
-            return x86_64::xmm25;
-        case X86_REG_XMM26:
-            return x86_64::xmm26;
-        case X86_REG_XMM27:
-            return x86_64::xmm27;
-        case X86_REG_XMM28:
-            return x86_64::xmm28;
-        case X86_REG_XMM29:
-            return x86_64::xmm29;
-        case X86_REG_XMM30:
-            return x86_64::xmm30;
-        case X86_REG_XMM31:
-            return x86_64::xmm31;
-        case X86_REG_YMM0:
-            return x86_64::ymm0;
-        case X86_REG_YMM1:
-            return x86_64::ymm1;
-        case X86_REG_YMM2:
-            return x86_64::ymm2;
-        case X86_REG_YMM3:
-            return x86_64::ymm3;
-        case X86_REG_YMM4:
-            return x86_64::ymm4;
-        case X86_REG_YMM5:
-            return x86_64::ymm5;
-        case X86_REG_YMM6:
-            return x86_64::ymm6;
-        case X86_REG_YMM7:
-            return x86_64::ymm7;
-        case X86_REG_YMM8:
-            return x86_64::ymm8;
-        case X86_REG_YMM9:
-            return x86_64::ymm9;
-        case X86_REG_YMM10:
-            return x86_64::ymm10;
-        case X86_REG_YMM11:
-            return x86_64::ymm11;
-        case X86_REG_YMM12:
-            return x86_64::ymm12;
-        case X86_REG_YMM13:
-            return x86_64::ymm13;
-        case X86_REG_YMM14:
-            return x86_64::ymm14;
-        case X86_REG_YMM15:
-            return x86_64::ymm15;
-        case X86_REG_YMM16:
-            return x86_64::ymm16;
-        case X86_REG_YMM17:
-            return x86_64::ymm17;
-        case X86_REG_YMM18:
-            return x86_64::ymm18;
-        case X86_REG_YMM19:
-            return x86_64::ymm19;
-        case X86_REG_YMM20:
-            return x86_64::ymm20;
-        case X86_REG_YMM21:
-            return x86_64::ymm21;
-        case X86_REG_YMM22:
-            return x86_64::ymm22;
-        case X86_REG_YMM23:
-            return x86_64::ymm23;
-        case X86_REG_YMM24:
-            return x86_64::ymm24;
-        case X86_REG_YMM25:
-            return x86_64::ymm25;
-        case X86_REG_YMM26:
-            return x86_64::ymm26;
-        case X86_REG_YMM27:
-            return x86_64::ymm27;
-        case X86_REG_YMM28:
-            return x86_64::ymm28;
-        case X86_REG_YMM29:
-            return x86_64::ymm29;
-        case X86_REG_YMM30:
-            return x86_64::ymm30;
-        case X86_REG_YMM31:
-            return x86_64::ymm31;
-        case X86_REG_ZMM0:
-            return x86_64::zmm0;
-        case X86_REG_ZMM1:
-            return x86_64::zmm1;
-        case X86_REG_ZMM2:
-            return x86_64::zmm2;
-        case X86_REG_ZMM3:
-            return x86_64::zmm3;
-        case X86_REG_ZMM4:
-            return x86_64::zmm4;
-        case X86_REG_ZMM5:
-            return x86_64::zmm5;
-        case X86_REG_ZMM6:
-            return x86_64::zmm6;
-        case X86_REG_ZMM7:
-            return x86_64::zmm7;
-        case X86_REG_ZMM8:
-            return x86_64::zmm8;
-        case X86_REG_ZMM9:
-            return x86_64::zmm9;
-        case X86_REG_ZMM10:
-            return x86_64::zmm10;
-        case X86_REG_ZMM11:
-            return x86_64::zmm11;
-        case X86_REG_ZMM12:
-            return x86_64::zmm12;
-        case X86_REG_ZMM13:
-            return x86_64::zmm13;
-        case X86_REG_ZMM14:
-            return x86_64::zmm14;
-        case X86_REG_ZMM15:
-            return x86_64::zmm15;
-        case X86_REG_ZMM16:
-            return x86_64::zmm16;
-        case X86_REG_ZMM17:
-            return x86_64::zmm17;
-        case X86_REG_ZMM18:
-            return x86_64::zmm18;
-        case X86_REG_ZMM19:
-            return x86_64::zmm19;
-        case X86_REG_ZMM20:
-            return x86_64::zmm20;
-        case X86_REG_ZMM21:
-            return x86_64::zmm21;
-        case X86_REG_ZMM22:
-            return x86_64::zmm22;
-        case X86_REG_ZMM23:
-            return x86_64::zmm23;
-        case X86_REG_ZMM24:
-            return x86_64::zmm24;
-        case X86_REG_ZMM25:
-            return x86_64::zmm25;
-        case X86_REG_ZMM26:
-            return x86_64::zmm26;
-        case X86_REG_ZMM27:
-            return x86_64::zmm27;
-        case X86_REG_ZMM28:
-            return x86_64::zmm28;
-        case X86_REG_ZMM29:
-            return x86_64::zmm29;
-        case X86_REG_ZMM30:
-            return x86_64::zmm30;
-        case X86_REG_ZMM31:
-            return x86_64::zmm31;
-        case X86_REG_R8B:
-            return x86_64::r8b;
-        case X86_REG_R9B:
-            return x86_64::r9b;
-        case X86_REG_R10B:
-            return x86_64::r10b;
-        case X86_REG_R11B:
-            return x86_64::r11b;
-        case X86_REG_R12B:
-            return x86_64::r12b;
-        case X86_REG_R13B:
-            return x86_64::r13b;
-        case X86_REG_R14B:
-            return x86_64::r14b;
-        case X86_REG_R15B:
-            return x86_64::r15b;
-        case X86_REG_R8D:
-            return x86_64::r8d;
-        case X86_REG_R9D:
-            return x86_64::r9d;
-        case X86_REG_R10D:
-            return x86_64::r10d;
-        case X86_REG_R11D:
-            return x86_64::r11d;
-        case X86_REG_R12D:
-            return x86_64::r12d;
-        case X86_REG_R13D:
-            return x86_64::r13d;
-        case X86_REG_R14D:
-            return x86_64::r14d;
-        case X86_REG_R15D:
-            return x86_64::r15d;
-        case X86_REG_R8W:
-            return x86_64::r8w;
-        case X86_REG_R9W:
-            return x86_64::r9w;
-        case X86_REG_R10W:
-            return x86_64::r10w;
-        case X86_REG_R11W:
-            return x86_64::r11w;
-        case X86_REG_R12W:
-            return x86_64::r12w;
-        case X86_REG_R13W:
-            return x86_64::r13w;
-        case X86_REG_R14W:
-            return x86_64::r14w;
-        case X86_REG_R15W:
-            return x86_64::r15w;
-            /*
-        case X86_REG_BND0:
-            return x86_64::bnd0;
-        case X86_REG_BND1:
-            return x86_64::bnd1;
-        case X86_REG_BND2:
-            return x86_64::bnd2;
-        case X86_REG_BND3:
-            return x86_64::bnd3;
-            */
-        default:
-            return InvalidReg;
-    }
-}
-
-
-
 /******************************  ppc functions  *************************************/
 
 void InstructionDecoder_Capstone::decodeOperands_ppc(const Instruction* insn, cs_detail *d) {
